--- public/assets/index-Bh74FOVn.js
@@ -1,142 +1,0 @@
-var un=Object.defineProperty;var hn=(t,e,n)=>e in t?un(t,e,{enumerable:!0,configurable:!0,writable:!0,value:n}):t[e]=n;var ht=(t,e,n)=>hn(t,typeof e!="symbol"?e+"":e,n);(function(){const e=document.createElement("link").relList;if(e&&e.supports&&e.supports("modulepreload"))return;for(const r of document.querySelectorAll('link[rel="modulepreload"]'))i(r);new MutationObserver(r=>{for(const s of r)if(s.type==="childList")for(const o of s.addedNodes)o.tagName==="LINK"&&o.rel==="modulepreload"&&i(o)}).observe(document,{childList:!0,subtree:!0});function n(r){const s={};return r.integrity&&(s.integrity=r.integrity),r.referrerPolicy&&(s.referrerPolicy=r.referrerPolicy),r.crossOrigin==="use-credentials"?s.credentials="include":r.crossOrigin==="anonymous"?s.credentials="omit":s.credentials="same-origin",s}function i(r){if(r.ep)return;r.ep=!0;const s=n(r);fetch(r.href,s)}})();function U(t=0){var n,i;const e=document.getElementById("loading");e&&(e.classList.remove("hidden"),document.getElementById("loadProgress").value=t,(n=document.getElementById("error"))==null||n.classList.add("hidden"),(i=document.getElementById("feedback"))==null||i.classList.add("hidden"))}function K(){const t=document.getElementById("loading");t&&t.classList.add("hidden")}function Rt(t){var n;K();const e=document.getElementById("feedback");e&&(e.textContent=t,e.classList.remove("hidden"),(n=document.getElementById("error"))==null||n.classList.add("hidden"))}function et(t){var n;K();const e=document.getElementById("error");e&&(e.textContent=t,e.classList.remove("hidden"),(n=document.getElementById("feedback"))==null||n.classList.add("hidden"))}function Jt(t,e){let n;return(...i)=>{clearTimeout(n),n=setTimeout(()=>t(...i),e)}}class fn{constructor(e,n={}){ht(this,"exportHTML",Jt(async()=>{try{const e=`
-        <!DOCTYPE html>
-        <html><body><h1>Sitemap URLs</h1><ul>
-        ${this.options.urls.map(n=>`<li><a href="${n.loc}">${n.loc}</a> (Last Modified: ${n.lastmod})</li>`).join("")}
-        </ul></body></html>`;this.downloadFile(e,"sitemap.html","text/html")}catch{throw new Error("HTML export failed")}},300));ht(this,"exportCSV",Jt(async()=>{try{const e=`URL,Last Modified,Change Frequency,Priority
-`+this.options.urls.map(n=>`"${n.loc}","${n.lastmod}","${n.changefreq}","${n.priority}"`).join(`
-`);this.downloadFile(e,"sitemap.csv","text/csv")}catch{throw new Error("CSV export failed")}},300));ht(this,"exportJSON",Jt(async()=>{try{const e=JSON.stringify(this.options.urls,null,2);this.downloadFile(e,"sitemap.json","application/json")}catch{throw new Error("JSON export failed")}},300));this.container=e,this.options={urls:n.urls||[],path:n.path||window.location.pathname,...n},this.listeners=new Map,this.render(),this.setupEventListeners()}render(){const e=this.options.path==="/results";this.container.innerHTML=`
-      <button id="exportHtmlBtn" ${!e||!this.options.urls.length?"hidden":""} aria-label="Export as HTML">
-        <i class="fas fa-file-code"></i> Export HTML
-      </button>
-      <button id="exportCsvBtn" ${!e||!this.options.urls.length?"hidden":""} aria-label="Export as CSV">
-        <i class="fas fa-file-csv"></i> Export CSV
-      </button>
-      <button id="exportJsonBtn" ${!e||!this.options.urls.length?"hidden":""} aria-label="Export as JSON">
-        <i class="fas fa-file-code"></i> Export JSON
-      </button>
-      <button id="clearCacheBtn" ${!e||!this.options.urls.length?"":"hidden"} aria-label="Clear cache">
-        <i class="fas fa-trash"></i> Clear Cache
-      </button>
-     
-      <button id="themeToggleBtn" aria-label="Toggle dark mode">
-        <i class="fas fa-moon"></i>
-      </button>
-    `}setupEventListeners(){this.cleanupListeners(),[["#exportHtmlBtn",this.exportHTML.bind(this),"Exporting HTML..."],["#exportCsvBtn",this.exportCSV.bind(this),"Exporting CSV..."],["#exportJsonBtn",this.exportJSON.bind(this),"Exporting JSON..."]].forEach(([r,s,o])=>{const a=this.container.querySelector(r);if(a){const u=async()=>{a.disabled=!0,a.textContent=o;try{await s()}catch(c){console.error(`Export failed (${r}):`,c),alert(`Failed to export: ${c.message}`)}a.disabled=!this.options.urls.length||this.options.path!=="/results",a.innerHTML=a.id==="exportHtmlBtn"?'<i class="fas fa-file-code"></i> Export HTML':a.id==="exportCsvBtn"?'<i class="fas fa-file-csv"></i> Export CSV':'<i class="fas fa-file-code"></i> Export JSON'};a.addEventListener("click",u),this.listeners.set(a,[["click",u]])}});const n=this.container.querySelector("#clearCacheBtn");if(n){const r=async()=>{U();try{if(!(await fetch("/sitemap/clear-cache",{method:"POST"})).ok)throw new Error("Failed to clear cache");Rt("Cache cleared successfully"),setTimeout(()=>window.location.reload(),1e3)}catch(s){et(s.message)}finally{K()}};n.addEventListener("click",r),this.listeners.set(n,[["click",r]])}const i=this.container.querySelector("#themeToggleBtn");if(i){const r=()=>document.body.classList.toggle("dark-mode");i.addEventListener("click",r),this.listeners.set(i,[["click",r]])}}cleanupListeners(){this.listeners.forEach((e,n)=>{e.forEach(([i,r])=>n.removeEventListener(i,r))}),this.listeners.clear()}downloadFile(e,n,i){try{const r=new Blob([e],{type:i}),s=document.createElement("a");s.href=URL.createObjectURL(r),s.download=n,document.body.appendChild(s),s.click(),document.body.removeChild(s),URL.revokeObjectURL(s.href)}catch(r){throw new Error(`Failed to download ${n}: ${r.message}`)}}updateRoute(e){this.options.path=e,this.render(),this.setupEventListeners()}updateUrls(e){this.options.urls=e,this.render(),this.setupEventListeners()}updateOptions(e){this.options={...this.options,...e},this.render(),this.setupEventListeners()}destroy(){this.cleanupListeners(),this.container.innerHTML=""}}class dn{constructor(e,n={}){this.container=e,this.options={onBack:n.onBack||(()=>{}),onToggleFilters:n.onToggleFilters||(()=>{}),treeView:n.treeView||null,urls:n.urls||[],destroy:n.destroy||(()=>{}),...n},this.path=window.location.pathname,this.buttons=null,this.render()}render(){this.container.innerHTML=`
-      <h1>Sitemap Explorer</h1>
-      <div class="controls"></div>
-    `;const e=this.container.querySelector(".controls");this.buttons=new fn(e,{onBack:this.options.onBack,urls:this.options.urls,path:this.path})}updateRoute(e){this.path=e,this.buttons.updateRoute(e)}updateUrls(e){this.options.urls=e,this.buttons.updateUrls(e)}updateOptions(e){this.options.destroy&&this.options.destroy(),this.options={...this.options,...e},this.buttons.updateOptions({onBack:this.options.onBack,urls:this.options.urls}),this.render()}destroy(){this.buttons&&this.buttons.destroy(),this.options.destroy&&this.options.destroy(),this.container.innerHTML=""}}function pn(t){try{const e=new URL(t);return e.protocol==="http:"||e.protocol==="https:"}catch{return!1}}function mn(t){t.innerHTML=`
-    <h2>Load Sitemap</h2>
-    <form id="sitemapUrlForm" method="POST" class="input-group">
-      <label for="sitemapUrlInput">Enter Sitemap URL:</label>
-      <input type="text" id="sitemapUrlInput" name="url" placeholder="https://example.com/sitemap.xml" />
-      <button type="submit">Load Sitemap</button>
-    </form>
-    <form id="sitemapUploadForm" method="POST" enctype="multipart/form-data" class="input-group">
-      <label for="sitemapFile">Upload Sitemap File:</label>
-      <input type="file" id="sitemapFile" name="sitemapFile" accept=".xml" />
-      <button type="submit">Upload</button>
-    </form>
-    <div id="status">
-      <div id="loading" class="hidden">
-        <progress id="loadProgress" max="100" value="0"></progress>
-        <span>Loading...</span>
-      </div>
-      <div id="feedback" class="hidden"></div>
-      <div id="error" class="hidden"></div>
-    </div>
-  `;const e=t.querySelector("#sitemapUrlForm"),n=t.querySelector("#sitemapUrlInput"),i=t.querySelector("#sitemapUploadForm");return e.addEventListener("submit",async r=>{r.preventDefault();const s=n.value.trim();if(!s){et("Please enter a sitemap URL");return}if(!pn(s)){et("Invalid URL: Must start with http:// or https:// and be well-formed");return}U(10);try{const o=await fetch("/sitemap/url",{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify({url:s})});if(!o.ok){const l=await o.json();throw new Error(l.error||"Failed to fetch sitemap")}const{id:a,message:u,status:c}=await o.json();U(50),window.history.pushState({},"",`/results?id=${a}`),Ot("/results"),Rt(`${u} (${c.succeeded.length}/${c.attempted.length} sitemaps succeeded)`)}catch(o){et(o.message)}finally{K()}}),i.addEventListener("submit",async r=>{r.preventDefault();const s=new FormData(i);U(10);try{const o=await fetch("/sitemap/upload",{method:"POST",body:s});if(!o.ok){const c=await o.json();throw new Error(c.error||"Failed to upload sitemap")}const{id:a,message:u}=await o.json();U(50),window.history.pushState({},"",`/results?id=${a}`),Ot("/results"),Rt(u)}catch(o){et(o.message)}finally{K()}}),{}}function gn(t,e){let n;return(...i)=>{clearTimeout(n),n=setTimeout(()=>t(...i),e)}}class yn{constructor(e){ht(this,"render",gn((e,n={},i="url",r="")=>{U(50),this.urls=e;const s=this.filterUrls(e,n,r);this.treeData=this.buildTree(s,i),this.container.querySelector(".tree-list")||(this.container.innerHTML="",this.renderControls());const o=this.container.querySelector(".tree-list")||document.createElement("ul");o.className="tree-list",o.innerHTML="",this.renderTreeNode(this.treeData,o,"",i),this.container.contains(o)||this.container.appendChild(o),this.setupNodeListeners(),this.updateTreeVisibility(),K()},200));this.container=e,this.isExpanded=!0,this.expandedNodes=new Set,this.treeData=null,this.urls=[],this.listeners=new Map,this.renderControls()}renderControls(){const e=document.createElement("div");e.className="tree-controls",e.innerHTML=`
-      <button id="toggleExpansionBtn" aria-expanded="${this.isExpanded}" aria-label="${this.isExpanded?"Collapse all nodes":"Expand all nodes"}">
-        <i class="fas fa-${this.isExpanded?"compress-alt":"expand-alt"}"></i> ${this.isExpanded?"Collapse All":"Expand All"}
-      </button>
-    `,this.container.appendChild(e),this.setupControlListeners()}setupControlListeners(){this.cleanupListeners("controls");const e=this.container.querySelector("#toggleExpansionBtn");if(e){const n=()=>this.toggleExpansion();e.addEventListener("click",n),this.listeners.set(e,[["click",n]])}}toggleExpansion(){this.isExpanded=!this.isExpanded;const e=this.container.querySelector("#toggleExpansionBtn");e.innerHTML=`<i class="fas fa-${this.isExpanded?"compress-alt":"expand-alt"}"></i> ${this.isExpanded?"Collapse All":"Expand All"}`,e.setAttribute("aria-expanded",this.isExpanded),e.setAttribute("aria-label",this.isExpanded?"Collapse all nodes":"Expand all nodes"),this.expandedNodes.clear(),this.isExpanded&&this.buildExpandedNodes(this.treeData,""),this.updateTreeVisibility()}buildExpandedNodes(e,n){Object.entries(e.children).forEach(([i,r])=>{const s=n?`${n}/${r.name}`:r.name;this.expandedNodes.add(s),this.buildExpandedNodes(r,s)})}updateTreeVisibility(){this.container.querySelectorAll(".toggle").forEach(e=>{const n=e.dataset.path,i=this.isExpanded||this.expandedNodes.has(n);e.classList.toggle("open",i),e.setAttribute("aria-expanded",i);const r=e.nextElementSibling;r?(r.classList.toggle("hidden",!i),r.style.display=i?"block":"none"):console.warn(`No child <ul> found for toggle at path: ${n}`)})}setupNodeListeners(){this.cleanupListeners("nodes"),this.container.querySelectorAll(".toggle").forEach(e=>{const n=()=>this.toggleNode(e);e.addEventListener("click",n),e.addEventListener("keydown",i=>i.key==="Enter"&&this.toggleNode(e)),this.listeners.set(e,[["click",n],["keydown",i=>i.key==="Enter"&&this.toggleNode(e)]])})}cleanupListeners(e="all"){this.listeners.forEach((n,i)=>{(e==="all"||e==="controls"&&i.id==="toggleExpansionBtn"||e==="nodes"&&i.classList.contains("toggle"))&&(n.forEach(([r,s])=>i.removeEventListener(r,s)),this.listeners.delete(i))})}filterUrls(e,n,i){return e.filter(r=>{const s=parseFloat(r.priority)||0,o=r.lastmod?new Date(r.lastmod):null,a=parseFloat(n.priority)||0,u=n.lastmod?new Date(n.lastmod):null;return(!i||r.loc.toLowerCase().includes(i.toLowerCase()))&&s>=a&&(!u||o&&o>=u)})}sortUrls(e,n){return e.sort((i,r)=>n==="lastmod"?new Date(r.lastmod||0)-new Date(i.lastmod||0)||i.loc.localeCompare(r.loc):n==="priority"&&parseFloat(r.priority||0)-parseFloat(i.priority||0)||i.loc.localeCompare(r.loc))}buildTree(e,n){const i={children:{},urls:[]};return e.forEach(r=>{try{const s=new URL(r.loc||""),o=s.pathname==="/"?[""]:s.pathname.split("/").filter(c=>c.length);let a=i;const u=s.hostname;a.children[u]||(a.children[u]={name:u,children:{},urls:[]}),a=a.children[u],o.forEach(c=>{const l=c||"/";a.children[l]||(a.children[l]={name:l,children:{},urls:[]}),a=a.children[l]}),a.urls.push(r)}catch(s){console.warn(`Skipping invalid URL: ${r.loc}`,s)}}),this.sortTree(i,n),i}sortTree(e,n){e.urls.length&&(e.urls=this.sortUrls(e.urls,n)),Object.values(e.children).forEach(i=>this.sortTree(i,n))}renderTreeNode(e,n,i,r){Object.entries(e.children).sort(([o,a],[u,c])=>r==="url"?a.name.localeCompare(c.name):0).forEach(([o,a])=>{const u=i?`${i}/${a.name}`:a.name,c=document.createElement("li"),l=document.createElement("span"),f=this.isExpanded||this.expandedNodes.has(u);if(l.className=`toggle ${f?"open":""}`,l.setAttribute("aria-expanded",f),l.setAttribute("aria-label",`Toggle ${a.name} children`),l.setAttribute("role","button"),l.dataset.path=u,l.tabIndex=0,l.textContent=a.name==="/"?u:a.name,c.appendChild(l),a.urls.length||Object.keys(a.children).length){const d=document.createElement("ul");d.className=f?"":"hidden",d.style.display=f?"block":"none",a.urls.forEach(m=>{const v=document.createElement("li");v.innerHTML=`<span class="url-leaf" tabindex="0" aria-label="URL: ${m.loc}">${this.formatUrl(m)}</span>`,d.appendChild(v)}),Object.keys(a.children).length&&this.renderTreeNode(a,d,u,r),c.appendChild(d)}n.appendChild(c)})}formatUrl(e){let n=e.loc;return e.lastmod&&(n+=` (Last Modified: ${e.lastmod})`),e.changefreq&&(n+=`, ${e.changefreq}`),e.priority&&(n+=`, Priority: ${e.priority}`),n}toggleNode(e){const i=!e.classList.contains("open");e.classList.toggle("open",i),e.setAttribute("aria-expanded",i);const r=e.nextElementSibling,s=e.dataset.path;if(r?(r.classList.toggle("hidden",!i),r.style.display=i?"block":"none",console.debug(`Toggled ${s}: ${i?"shown":"hidden"}`)):console.warn(`No child <ul> found for toggle at path: ${s}`),i?this.expandedNodes.add(s):this.expandedNodes.delete(s),this.isExpanded&&!i){this.isExpanded=!1;const o=this.container.querySelector("#toggleExpansionBtn");o.innerHTML='<i class="fas fa-expand-alt"></i> Expand All',o.setAttribute("aria-expanded",!1),o.setAttribute("aria-label","Expand all nodes")}}destroy(){this.cleanupListeners(),this.container.innerHTML="",this.treeData=null,this.urls=[],this.expandedNodes.clear()}}var wn={value:()=>{}};function ce(){for(var t=0,e=arguments.length,n={},i;t<e;++t){if(!(i=arguments[t]+"")||i in n||/[\s.]/.test(i))throw new Error("illegal type: "+i);n[i]=[]}return new Mt(n)}function Mt(t){this._=t}function vn(t,e){return t.trim().split(/^|\s+/).map(function(n){var i="",r=n.indexOf(".");if(r>=0&&(i=n.slice(r+1),n=n.slice(0,r)),n&&!e.hasOwnProperty(n))throw new Error("unknown type: "+n);return{type:n,name:i}})}Mt.prototype=ce.prototype={constructor:Mt,on:function(t,e){var n=this._,i=vn(t+"",n),r,s=-1,o=i.length;if(arguments.length<2){for(;++s<o;)if((r=(t=i[s]).type)&&(r=xn(n[r],t.name)))return r;return}if(e!=null&&typeof e!="function")throw new Error("invalid callback: "+e);for(;++s<o;)if(r=(t=i[s]).type)n[r]=ge(n[r],t.name,e);else if(e==null)for(r in n)n[r]=ge(n[r],t.name,null);return this},copy:function(){var t={},e=this._;for(var n in e)t[n]=e[n].slice();return new Mt(t)},call:function(t,e){if((r=arguments.length-2)>0)for(var n=new Array(r),i=0,r,s;i<r;++i)n[i]=arguments[i+2];if(!this._.hasOwnProperty(t))throw new Error("unknown type: "+t);for(s=this._[t],i=0,r=s.length;i<r;++i)s[i].value.apply(e,n)},apply:function(t,e,n){if(!this._.hasOwnProperty(t))throw new Error("unknown type: "+t);for(var i=this._[t],r=0,s=i.length;r<s;++r)i[r].value.apply(e,n)}};function xn(t,e){for(var n=0,i=t.length,r;n<i;++n)if((r=t[n]).name===e)return r.value}function ge(t,e,n){for(var i=0,r=t.length;i<r;++i)if(t[i].name===e){t[i]=wn,t=t.slice(0,i).concat(t.slice(i+1));break}return n!=null&&t.push({name:e,value:n}),t}var jt="http://www.w3.org/1999/xhtml";const ye={svg:"http://www.w3.org/2000/svg",xhtml:jt,xlink:"http://www.w3.org/1999/xlink",xml:"http://www.w3.org/XML/1998/namespace",xmlns:"http://www.w3.org/2000/xmlns/"};function Vt(t){var e=t+="",n=e.indexOf(":");return n>=0&&(e=t.slice(0,n))!=="xmlns"&&(t=t.slice(n+1)),ye.hasOwnProperty(e)?{space:ye[e],local:t}:t}function _n(t){return function(){var e=this.ownerDocument,n=this.namespaceURI;return n===jt&&e.documentElement.namespaceURI===jt?e.createElement(t):e.createElementNS(n,t)}}function bn(t){return function(){return this.ownerDocument.createElementNS(t.space,t.local)}}function Re(t){var e=Vt(t);return(e.local?bn:_n)(e)}function En(){}function ue(t){return t==null?En:function(){return this.querySelector(t)}}function Ln(t){typeof t!="function"&&(t=ue(t));for(var e=this._groups,n=e.length,i=new Array(n),r=0;r<n;++r)for(var s=e[r],o=s.length,a=i[r]=new Array(o),u,c,l=0;l<o;++l)(u=s[l])&&(c=t.call(u,u.__data__,l,s))&&("__data__"in u&&(c.__data__=u.__data__),a[l]=c);return new R(i,this._parents)}function Sn(t){return t==null?[]:Array.isArray(t)?t:Array.from(t)}function kn(){return[]}function Fe(t){return t==null?kn:function(){return this.querySelectorAll(t)}}function Cn(t){return function(){return Sn(t.apply(this,arguments))}}function $n(t){typeof t=="function"?t=Cn(t):t=Fe(t);for(var e=this._groups,n=e.length,i=[],r=[],s=0;s<n;++s)for(var o=e[s],a=o.length,u,c=0;c<a;++c)(u=o[c])&&(i.push(t.call(u,u.__data__,c,o)),r.push(u));return new R(i,r)}function He(t){return function(){return this.matches(t)}}function De(t){return function(e){return e.matches(t)}}var Tn=Array.prototype.find;function Mn(t){return function(){return Tn.call(this.children,t)}}function zn(){return this.firstElementChild}function Nn(t){return this.select(t==null?zn:Mn(typeof t=="function"?t:De(t)))}var An=Array.prototype.filter;function Bn(){return Array.from(this.children)}function Rn(t){return function(){return An.call(this.children,t)}}function Fn(t){return this.selectAll(t==null?Bn:Rn(typeof t=="function"?t:De(t)))}function Hn(t){typeof t!="function"&&(t=He(t));for(var e=this._groups,n=e.length,i=new Array(n),r=0;r<n;++r)for(var s=e[r],o=s.length,a=i[r]=[],u,c=0;c<o;++c)(u=s[c])&&t.call(u,u.__data__,c,s)&&a.push(u);return new R(i,this._parents)}function Ue(t){return new Array(t.length)}function Dn(){return new R(this._enter||this._groups.map(Ue),this._parents)}function Ft(t,e){this.ownerDocument=t.ownerDocument,this.namespaceURI=t.namespaceURI,this._next=null,this._parent=t,this.__data__=e}Ft.prototype={constructor:Ft,appendChild:function(t){return this._parent.insertBefore(t,this._next)},insertBefore:function(t,e){return this._parent.insertBefore(t,e)},querySelector:function(t){return this._parent.querySelector(t)},querySelectorAll:function(t){return this._parent.querySelectorAll(t)}};function Un(t){return function(){return t}}function qn(t,e,n,i,r,s){for(var o=0,a,u=e.length,c=s.length;o<c;++o)(a=e[o])?(a.__data__=s[o],i[o]=a):n[o]=new Ft(t,s[o]);for(;o<u;++o)(a=e[o])&&(r[o]=a)}function In(t,e,n,i,r,s,o){var a,u,c=new Map,l=e.length,f=s.length,d=new Array(l),m;for(a=0;a<l;++a)(u=e[a])&&(d[a]=m=o.call(u,u.__data__,a,e)+"",c.has(m)?r[a]=u:c.set(m,u));for(a=0;a<f;++a)m=o.call(t,s[a],a,s)+"",(u=c.get(m))?(i[a]=u,u.__data__=s[a],c.delete(m)):n[a]=new Ft(t,s[a]);for(a=0;a<l;++a)(u=e[a])&&c.get(d[a])===u&&(r[a]=u)}function On(t){return t.__data__}function Vn(t,e){if(!arguments.length)return Array.from(this,On);var n=e?In:qn,i=this._parents,r=this._groups;typeof t!="function"&&(t=Un(t));for(var s=r.length,o=new Array(s),a=new Array(s),u=new Array(s),c=0;c<s;++c){var l=i[c],f=r[c],d=f.length,m=Pn(t.call(l,l&&l.__data__,c,i)),v=m.length,_=a[c]=new Array(v),S=o[c]=new Array(v),w=u[c]=new Array(d);n(l,f,_,S,w,m,e);for(var b=0,C=0,$,z;b<v;++b)if($=_[b]){for(b>=C&&(C=b+1);!(z=S[C])&&++C<v;);$._next=z||null}}return o=new R(o,i),o._enter=a,o._exit=u,o}function Pn(t){return typeof t=="object"&&"length"in t?t:Array.from(t)}function Xn(){return new R(this._exit||this._groups.map(Ue),this._parents)}function Yn(t,e,n){var i=this.enter(),r=this,s=this.exit();return typeof t=="function"?(i=t(i),i&&(i=i.selection())):i=i.append(t+""),e!=null&&(r=e(r),r&&(r=r.selection())),n==null?s.remove():n(s),i&&r?i.merge(r).order():r}function Jn(t){for(var e=t.selection?t.selection():t,n=this._groups,i=e._groups,r=n.length,s=i.length,o=Math.min(r,s),a=new Array(r),u=0;u<o;++u)for(var c=n[u],l=i[u],f=c.length,d=a[u]=new Array(f),m,v=0;v<f;++v)(m=c[v]||l[v])&&(d[v]=m);for(;u<r;++u)a[u]=n[u];return new R(a,this._parents)}function Kn(){for(var t=this._groups,e=-1,n=t.length;++e<n;)for(var i=t[e],r=i.length-1,s=i[r],o;--r>=0;)(o=i[r])&&(s&&o.compareDocumentPosition(s)^4&&s.parentNode.insertBefore(o,s),s=o);return this}function Zn(t){t||(t=Gn);function e(f,d){return f&&d?t(f.__data__,d.__data__):!f-!d}for(var n=this._groups,i=n.length,r=new Array(i),s=0;s<i;++s){for(var o=n[s],a=o.length,u=r[s]=new Array(a),c,l=0;l<a;++l)(c=o[l])&&(u[l]=c);u.sort(e)}return new R(r,this._parents).order()}function Gn(t,e){return t<e?-1:t>e?1:t>=e?0:NaN}function Wn(){var t=arguments[0];return arguments[0]=this,t.apply(null,arguments),this}function Qn(){return Array.from(this)}function jn(){for(var t=this._groups,e=0,n=t.length;e<n;++e)for(var i=t[e],r=0,s=i.length;r<s;++r){var o=i[r];if(o)return o}return null}function ti(){let t=0;for(const e of this)++t;return t}function ei(){return!this.node()}function ni(t){for(var e=this._groups,n=0,i=e.length;n<i;++n)for(var r=e[n],s=0,o=r.length,a;s<o;++s)(a=r[s])&&t.call(a,a.__data__,s,r);return this}function ii(t){return function(){this.removeAttribute(t)}}function ri(t){return function(){this.removeAttributeNS(t.space,t.local)}}function si(t,e){return function(){this.setAttribute(t,e)}}function oi(t,e){return function(){this.setAttributeNS(t.space,t.local,e)}}function ai(t,e){return function(){var n=e.apply(this,arguments);n==null?this.removeAttribute(t):this.setAttribute(t,n)}}function li(t,e){return function(){var n=e.apply(this,arguments);n==null?this.removeAttributeNS(t.space,t.local):this.setAttributeNS(t.space,t.local,n)}}function ci(t,e){var n=Vt(t);if(arguments.length<2){var i=this.node();return n.local?i.getAttributeNS(n.space,n.local):i.getAttribute(n)}return this.each((e==null?n.local?ri:ii:typeof e=="function"?n.local?li:ai:n.local?oi:si)(n,e))}function qe(t){return t.ownerDocument&&t.ownerDocument.defaultView||t.document&&t||t.defaultView}function ui(t){return function(){this.style.removeProperty(t)}}function hi(t,e,n){return function(){this.style.setProperty(t,e,n)}}function fi(t,e,n){return function(){var i=e.apply(this,arguments);i==null?this.style.removeProperty(t):this.style.setProperty(t,i,n)}}function di(t,e,n){return arguments.length>1?this.each((e==null?ui:typeof e=="function"?fi:hi)(t,e,n??"")):ct(this.node(),t)}function ct(t,e){return t.style.getPropertyValue(e)||qe(t).getComputedStyle(t,null).getPropertyValue(e)}function pi(t){return function(){delete this[t]}}function mi(t,e){return function(){this[t]=e}}function gi(t,e){return function(){var n=e.apply(this,arguments);n==null?delete this[t]:this[t]=n}}function yi(t,e){return arguments.length>1?this.each((e==null?pi:typeof e=="function"?gi:mi)(t,e)):this.node()[t]}function Ie(t){return t.trim().split(/^|\s+/)}function he(t){return t.classList||new Oe(t)}function Oe(t){this._node=t,this._names=Ie(t.getAttribute("class")||"")}Oe.prototype={add:function(t){var e=this._names.indexOf(t);e<0&&(this._names.push(t),this._node.setAttribute("class",this._names.join(" ")))},remove:function(t){var e=this._names.indexOf(t);e>=0&&(this._names.splice(e,1),this._node.setAttribute("class",this._names.join(" ")))},contains:function(t){return this._names.indexOf(t)>=0}};function Ve(t,e){for(var n=he(t),i=-1,r=e.length;++i<r;)n.add(e[i])}function Pe(t,e){for(var n=he(t),i=-1,r=e.length;++i<r;)n.remove(e[i])}function wi(t){return function(){Ve(this,t)}}function vi(t){return function(){Pe(this,t)}}function xi(t,e){return function(){(e.apply(this,arguments)?Ve:Pe)(this,t)}}function _i(t,e){var n=Ie(t+"");if(arguments.length<2){for(var i=he(this.node()),r=-1,s=n.length;++r<s;)if(!i.contains(n[r]))return!1;return!0}return this.each((typeof e=="function"?xi:e?wi:vi)(n,e))}function bi(){this.textContent=""}function Ei(t){return function(){this.textContent=t}}function Li(t){return function(){var e=t.apply(this,arguments);this.textContent=e??""}}function Si(t){return arguments.length?this.each(t==null?bi:(typeof t=="function"?Li:Ei)(t)):this.node().textContent}function ki(){this.innerHTML=""}function Ci(t){return function(){this.innerHTML=t}}function $i(t){return function(){var e=t.apply(this,arguments);this.innerHTML=e??""}}function Ti(t){return arguments.length?this.each(t==null?ki:(typeof t=="function"?$i:Ci)(t)):this.node().innerHTML}function Mi(){this.nextSibling&&this.parentNode.appendChild(this)}function zi(){return this.each(Mi)}function Ni(){this.previousSibling&&this.parentNode.insertBefore(this,this.parentNode.firstChild)}function Ai(){return this.each(Ni)}function Bi(t){var e=typeof t=="function"?t:Re(t);return this.select(function(){return this.appendChild(e.apply(this,arguments))})}function Ri(){return null}function Fi(t,e){var n=typeof t=="function"?t:Re(t),i=e==null?Ri:typeof e=="function"?e:ue(e);return this.select(function(){return this.insertBefore(n.apply(this,arguments),i.apply(this,arguments)||null)})}function Hi(){var t=this.parentNode;t&&t.removeChild(this)}function Di(){return this.each(Hi)}function Ui(){var t=this.cloneNode(!1),e=this.parentNode;return e?e.insertBefore(t,this.nextSibling):t}function qi(){var t=this.cloneNode(!0),e=this.parentNode;return e?e.insertBefore(t,this.nextSibling):t}function Ii(t){return this.select(t?qi:Ui)}function Oi(t){return arguments.length?this.property("__data__",t):this.node().__data__}function Vi(t){return function(e){t.call(this,e,this.__data__)}}function Pi(t){return t.trim().split(/^|\s+/).map(function(e){var n="",i=e.indexOf(".");return i>=0&&(n=e.slice(i+1),e=e.slice(0,i)),{type:e,name:n}})}function Xi(t){return function(){var e=this.__on;if(e){for(var n=0,i=-1,r=e.length,s;n<r;++n)s=e[n],(!t.type||s.type===t.type)&&s.name===t.name?this.removeEventListener(s.type,s.listener,s.options):e[++i]=s;++i?e.length=i:delete this.__on}}}function Yi(t,e,n){return function(){var i=this.__on,r,s=Vi(e);if(i){for(var o=0,a=i.length;o<a;++o)if((r=i[o]).type===t.type&&r.name===t.name){this.removeEventListener(r.type,r.listener,r.options),this.addEventListener(r.type,r.listener=s,r.options=n),r.value=e;return}}this.addEventListener(t.type,s,n),r={type:t.type,name:t.name,value:e,listener:s,options:n},i?i.push(r):this.__on=[r]}}function Ji(t,e,n){var i=Pi(t+""),r,s=i.length,o;if(arguments.length<2){var a=this.node().__on;if(a){for(var u=0,c=a.length,l;u<c;++u)for(r=0,l=a[u];r<s;++r)if((o=i[r]).type===l.type&&o.name===l.name)return l.value}return}for(a=e?Yi:Xi,r=0;r<s;++r)this.each(a(i[r],e,n));return this}function Xe(t,e,n){var i=qe(t),r=i.CustomEvent;typeof r=="function"?r=new r(e,n):(r=i.document.createEvent("Event"),n?(r.initEvent(e,n.bubbles,n.cancelable),r.detail=n.detail):r.initEvent(e,!1,!1)),t.dispatchEvent(r)}function Ki(t,e){return function(){return Xe(this,t,e)}}function Zi(t,e){return function(){return Xe(this,t,e.apply(this,arguments))}}function Gi(t,e){return this.each((typeof e=="function"?Zi:Ki)(t,e))}function*Wi(){for(var t=this._groups,e=0,n=t.length;e<n;++e)for(var i=t[e],r=0,s=i.length,o;r<s;++r)(o=i[r])&&(yield o)}var Ye=[null];function R(t,e){this._groups=t,this._parents=e}function _t(){return new R([[document.documentElement]],Ye)}function Qi(){return this}R.prototype=_t.prototype={constructor:R,select:Ln,selectAll:$n,selectChild:Nn,selectChildren:Fn,filter:Hn,data:Vn,enter:Dn,exit:Xn,join:Yn,merge:Jn,selection:Qi,order:Kn,sort:Zn,call:Wn,nodes:Qn,node:jn,size:ti,empty:ei,each:ni,attr:ci,style:di,property:yi,classed:_i,text:Si,html:Ti,raise:zi,lower:Ai,append:Bi,insert:Fi,remove:Di,clone:Ii,datum:Oi,on:Ji,dispatch:Gi,[Symbol.iterator]:Wi};function Q(t){return typeof t=="string"?new R([[document.querySelector(t)]],[document.documentElement]):new R([[t]],Ye)}function ji(t){let e;for(;e=t.sourceEvent;)t=e;return t}function j(t,e){if(t=ji(t),e===void 0&&(e=t.currentTarget),e){var n=e.ownerSVGElement||e;if(n.createSVGPoint){var i=n.createSVGPoint();return i.x=t.clientX,i.y=t.clientY,i=i.matrixTransform(e.getScreenCTM().inverse()),[i.x,i.y]}if(e.getBoundingClientRect){var r=e.getBoundingClientRect();return[t.clientX-r.left-e.clientLeft,t.clientY-r.top-e.clientTop]}}return[t.pageX,t.pageY]}const te={capture:!0,passive:!1};function ee(t){t.preventDefault(),t.stopImmediatePropagation()}function tr(t){var e=t.document.documentElement,n=Q(t).on("dragstart.drag",ee,te);"onselectstart"in e?n.on("selectstart.drag",ee,te):(e.__noselect=e.style.MozUserSelect,e.style.MozUserSelect="none")}function er(t,e){var n=t.document.documentElement,i=Q(t).on("dragstart.drag",null);e&&(i.on("click.drag",ee,te),setTimeout(function(){i.on("click.drag",null)},0)),"onselectstart"in n?i.on("selectstart.drag",null):(n.style.MozUserSelect=n.__noselect,delete n.__noselect)}function fe(t,e,n){t.prototype=e.prototype=n,n.constructor=t}function Je(t,e){var n=Object.create(t.prototype);for(var i in e)n[i]=e[i];return n}function bt(){}var gt=.7,Ht=1/gt,at="\\s*([+-]?\\d+)\\s*",yt="\\s*([+-]?(?:\\d*\\.)?\\d+(?:[eE][+-]?\\d+)?)\\s*",O="\\s*([+-]?(?:\\d*\\.)?\\d+(?:[eE][+-]?\\d+)?)%\\s*",nr=/^#([0-9a-f]{3,8})$/,ir=new RegExp(`^rgb\\(${at},${at},${at}\\)$`),rr=new RegExp(`^rgb\\(${O},${O},${O}\\)$`),sr=new RegExp(`^rgba\\(${at},${at},${at},${yt}\\)$`),or=new RegExp(`^rgba\\(${O},${O},${O},${yt}\\)$`),ar=new RegExp(`^hsl\\(${yt},${O},${O}\\)$`),lr=new RegExp(`^hsla\\(${yt},${O},${O},${yt}\\)$`),we={aliceblue:15792383,antiquewhite:16444375,aqua:65535,aquamarine:8388564,azure:15794175,beige:16119260,bisque:16770244,black:0,blanchedalmond:16772045,blue:255,blueviolet:9055202,brown:10824234,burlywood:14596231,cadetblue:6266528,chartreuse:8388352,chocolate:13789470,coral:16744272,cornflowerblue:6591981,cornsilk:16775388,crimson:14423100,cyan:65535,darkblue:139,darkcyan:35723,darkgoldenrod:12092939,darkgray:11119017,darkgreen:25600,darkgrey:11119017,darkkhaki:12433259,darkmagenta:9109643,darkolivegreen:5597999,darkorange:16747520,darkorchid:10040012,darkred:9109504,darksalmon:15308410,darkseagreen:9419919,darkslateblue:4734347,darkslategray:3100495,darkslategrey:3100495,darkturquoise:52945,darkviolet:9699539,deeppink:16716947,deepskyblue:49151,dimgray:6908265,dimgrey:6908265,dodgerblue:2003199,firebrick:11674146,floralwhite:16775920,forestgreen:2263842,fuchsia:16711935,gainsboro:14474460,ghostwhite:16316671,gold:16766720,goldenrod:14329120,gray:8421504,green:32768,greenyellow:11403055,grey:8421504,honeydew:15794160,hotpink:16738740,indianred:13458524,indigo:4915330,ivory:16777200,khaki:15787660,lavender:15132410,lavenderblush:16773365,lawngreen:8190976,lemonchiffon:16775885,lightblue:11393254,lightcoral:15761536,lightcyan:14745599,lightgoldenrodyellow:16448210,lightgray:13882323,lightgreen:9498256,lightgrey:13882323,lightpink:16758465,lightsalmon:16752762,lightseagreen:2142890,lightskyblue:8900346,lightslategray:7833753,lightslategrey:7833753,lightsteelblue:11584734,lightyellow:16777184,lime:65280,limegreen:3329330,linen:16445670,magenta:16711935,maroon:8388608,mediumaquamarine:6737322,mediumblue:205,mediumorchid:12211667,mediumpurple:9662683,mediumseagreen:3978097,mediumslateblue:8087790,mediumspringgreen:64154,mediumturquoise:4772300,mediumvioletred:13047173,midnightblue:1644912,mintcream:16121850,mistyrose:16770273,moccasin:16770229,navajowhite:16768685,navy:128,oldlace:16643558,olive:8421376,olivedrab:7048739,orange:16753920,orangered:16729344,orchid:14315734,palegoldenrod:15657130,palegreen:10025880,paleturquoise:11529966,palevioletred:14381203,papayawhip:16773077,peachpuff:16767673,peru:13468991,pink:16761035,plum:14524637,powderblue:11591910,purple:8388736,rebeccapurple:6697881,red:16711680,rosybrown:12357519,royalblue:4286945,saddlebrown:9127187,salmon:16416882,sandybrown:16032864,seagreen:3050327,seashell:16774638,sienna:10506797,silver:12632256,skyblue:8900331,slateblue:6970061,slategray:7372944,slategrey:7372944,snow:16775930,springgreen:65407,steelblue:4620980,tan:13808780,teal:32896,thistle:14204888,tomato:16737095,turquoise:4251856,violet:15631086,wheat:16113331,white:16777215,whitesmoke:16119285,yellow:16776960,yellowgreen:10145074};fe(bt,wt,{copy(t){return Object.assign(new this.constructor,this,t)},displayable(){return this.rgb().displayable()},hex:ve,formatHex:ve,formatHex8:cr,formatHsl:ur,formatRgb:xe,toString:xe});function ve(){return this.rgb().formatHex()}function cr(){return this.rgb().formatHex8()}function ur(){return Ke(this).formatHsl()}function xe(){return this.rgb().formatRgb()}function wt(t){var e,n;return t=(t+"").trim().toLowerCase(),(e=nr.exec(t))?(n=e[1].length,e=parseInt(e[1],16),n===6?_e(e):n===3?new A(e>>8&15|e>>4&240,e>>4&15|e&240,(e&15)<<4|e&15,1):n===8?kt(e>>24&255,e>>16&255,e>>8&255,(e&255)/255):n===4?kt(e>>12&15|e>>8&240,e>>8&15|e>>4&240,e>>4&15|e&240,((e&15)<<4|e&15)/255):null):(e=ir.exec(t))?new A(e[1],e[2],e[3],1):(e=rr.exec(t))?new A(e[1]*255/100,e[2]*255/100,e[3]*255/100,1):(e=sr.exec(t))?kt(e[1],e[2],e[3],e[4]):(e=or.exec(t))?kt(e[1]*255/100,e[2]*255/100,e[3]*255/100,e[4]):(e=ar.exec(t))?Le(e[1],e[2]/100,e[3]/100,1):(e=lr.exec(t))?Le(e[1],e[2]/100,e[3]/100,e[4]):we.hasOwnProperty(t)?_e(we[t]):t==="transparent"?new A(NaN,NaN,NaN,0):null}function _e(t){return new A(t>>16&255,t>>8&255,t&255,1)}function kt(t,e,n,i){return i<=0&&(t=e=n=NaN),new A(t,e,n,i)}function hr(t){return t instanceof bt||(t=wt(t)),t?(t=t.rgb(),new A(t.r,t.g,t.b,t.opacity)):new A}function ne(t,e,n,i){return arguments.length===1?hr(t):new A(t,e,n,i??1)}function A(t,e,n,i){this.r=+t,this.g=+e,this.b=+n,this.opacity=+i}fe(A,ne,Je(bt,{brighter(t){return t=t==null?Ht:Math.pow(Ht,t),new A(this.r*t,this.g*t,this.b*t,this.opacity)},darker(t){return t=t==null?gt:Math.pow(gt,t),new A(this.r*t,this.g*t,this.b*t,this.opacity)},rgb(){return this},clamp(){return new A(it(this.r),it(this.g),it(this.b),Dt(this.opacity))},displayable(){return-.5<=this.r&&this.r<255.5&&-.5<=this.g&&this.g<255.5&&-.5<=this.b&&this.b<255.5&&0<=this.opacity&&this.opacity<=1},hex:be,formatHex:be,formatHex8:fr,formatRgb:Ee,toString:Ee}));function be(){return`#${nt(this.r)}${nt(this.g)}${nt(this.b)}`}function fr(){return`#${nt(this.r)}${nt(this.g)}${nt(this.b)}${nt((isNaN(this.opacity)?1:this.opacity)*255)}`}function Ee(){const t=Dt(this.opacity);return`${t===1?"rgb(":"rgba("}${it(this.r)}, ${it(this.g)}, ${it(this.b)}${t===1?")":`, ${t})`}`}function Dt(t){return isNaN(t)?1:Math.max(0,Math.min(1,t))}function it(t){return Math.max(0,Math.min(255,Math.round(t)||0))}function nt(t){return t=it(t),(t<16?"0":"")+t.toString(16)}function Le(t,e,n,i){return i<=0?t=e=n=NaN:n<=0||n>=1?t=e=NaN:e<=0&&(t=NaN),new D(t,e,n,i)}function Ke(t){if(t instanceof D)return new D(t.h,t.s,t.l,t.opacity);if(t instanceof bt||(t=wt(t)),!t)return new D;if(t instanceof D)return t;t=t.rgb();var e=t.r/255,n=t.g/255,i=t.b/255,r=Math.min(e,n,i),s=Math.max(e,n,i),o=NaN,a=s-r,u=(s+r)/2;return a?(e===s?o=(n-i)/a+(n<i)*6:n===s?o=(i-e)/a+2:o=(e-n)/a+4,a/=u<.5?s+r:2-s-r,o*=60):a=u>0&&u<1?0:o,new D(o,a,u,t.opacity)}function dr(t,e,n,i){return arguments.length===1?Ke(t):new D(t,e,n,i??1)}function D(t,e,n,i){this.h=+t,this.s=+e,this.l=+n,this.opacity=+i}fe(D,dr,Je(bt,{brighter(t){return t=t==null?Ht:Math.pow(Ht,t),new D(this.h,this.s,this.l*t,this.opacity)},darker(t){return t=t==null?gt:Math.pow(gt,t),new D(this.h,this.s,this.l*t,this.opacity)},rgb(){var t=this.h%360+(this.h<0)*360,e=isNaN(t)||isNaN(this.s)?0:this.s,n=this.l,i=n+(n<.5?n:1-n)*e,r=2*n-i;return new A(Kt(t>=240?t-240:t+120,r,i),Kt(t,r,i),Kt(t<120?t+240:t-120,r,i),this.opacity)},clamp(){return new D(Se(this.h),Ct(this.s),Ct(this.l),Dt(this.opacity))},displayable(){return(0<=this.s&&this.s<=1||isNaN(this.s))&&0<=this.l&&this.l<=1&&0<=this.opacity&&this.opacity<=1},formatHsl(){const t=Dt(this.opacity);return`${t===1?"hsl(":"hsla("}${Se(this.h)}, ${Ct(this.s)*100}%, ${Ct(this.l)*100}%${t===1?")":`, ${t})`}`}}));function Se(t){return t=(t||0)%360,t<0?t+360:t}function Ct(t){return Math.max(0,Math.min(1,t||0))}function Kt(t,e,n){return(t<60?e+(n-e)*t/60:t<180?n:t<240?e+(n-e)*(240-t)/60:e)*255}const Ze=t=>()=>t;function pr(t,e){return function(n){return t+n*e}}function mr(t,e,n){return t=Math.pow(t,n),e=Math.pow(e,n)-t,n=1/n,function(i){return Math.pow(t+i*e,n)}}function gr(t){return(t=+t)==1?Ge:function(e,n){return n-e?mr(e,n,t):Ze(isNaN(e)?n:e)}}function Ge(t,e){var n=e-t;return n?pr(t,n):Ze(isNaN(t)?e:t)}const ke=function t(e){var n=gr(e);function i(r,s){var o=n((r=ne(r)).r,(s=ne(s)).r),a=n(r.g,s.g),u=n(r.b,s.b),c=Ge(r.opacity,s.opacity);return function(l){return r.r=o(l),r.g=a(l),r.b=u(l),r.opacity=c(l),r+""}}return i.gamma=t,i}(1);function W(t,e){return t=+t,e=+e,function(n){return t*(1-n)+e*n}}var ie=/[-+]?(?:\d+\.?\d*|\.?\d+)(?:[eE][-+]?\d+)?/g,Zt=new RegExp(ie.source,"g");function yr(t){return function(){return t}}function wr(t){return function(e){return t(e)+""}}function vr(t,e){var n=ie.lastIndex=Zt.lastIndex=0,i,r,s,o=-1,a=[],u=[];for(t=t+"",e=e+"";(i=ie.exec(t))&&(r=Zt.exec(e));)(s=r.index)>n&&(s=e.slice(n,s),a[o]?a[o]+=s:a[++o]=s),(i=i[0])===(r=r[0])?a[o]?a[o]+=r:a[++o]=r:(a[++o]=null,u.push({i:o,x:W(i,r)})),n=Zt.lastIndex;return n<e.length&&(s=e.slice(n),a[o]?a[o]+=s:a[++o]=s),a.length<2?u[0]?wr(u[0].x):yr(e):(e=u.length,function(c){for(var l=0,f;l<e;++l)a[(f=u[l]).i]=f.x(c);return a.join("")})}var Ce=180/Math.PI,re={translateX:0,translateY:0,rotate:0,skewX:0,scaleX:1,scaleY:1};function We(t,e,n,i,r,s){var o,a,u;return(o=Math.sqrt(t*t+e*e))&&(t/=o,e/=o),(u=t*n+e*i)&&(n-=t*u,i-=e*u),(a=Math.sqrt(n*n+i*i))&&(n/=a,i/=a,u/=a),t*i<e*n&&(t=-t,e=-e,u=-u,o=-o),{translateX:r,translateY:s,rotate:Math.atan2(e,t)*Ce,skewX:Math.atan(u)*Ce,scaleX:o,scaleY:a}}var $t;function xr(t){const e=new(typeof DOMMatrix=="function"?DOMMatrix:WebKitCSSMatrix)(t+"");return e.isIdentity?re:We(e.a,e.b,e.c,e.d,e.e,e.f)}function _r(t){return t==null||($t||($t=document.createElementNS("http://www.w3.org/2000/svg","g")),$t.setAttribute("transform",t),!(t=$t.transform.baseVal.consolidate()))?re:(t=t.matrix,We(t.a,t.b,t.c,t.d,t.e,t.f))}function Qe(t,e,n,i){function r(c){return c.length?c.pop()+" ":""}function s(c,l,f,d,m,v){if(c!==f||l!==d){var _=m.push("translate(",null,e,null,n);v.push({i:_-4,x:W(c,f)},{i:_-2,x:W(l,d)})}else(f||d)&&m.push("translate("+f+e+d+n)}function o(c,l,f,d){c!==l?(c-l>180?l+=360:l-c>180&&(c+=360),d.push({i:f.push(r(f)+"rotate(",null,i)-2,x:W(c,l)})):l&&f.push(r(f)+"rotate("+l+i)}function a(c,l,f,d){c!==l?d.push({i:f.push(r(f)+"skewX(",null,i)-2,x:W(c,l)}):l&&f.push(r(f)+"skewX("+l+i)}function u(c,l,f,d,m,v){if(c!==f||l!==d){var _=m.push(r(m)+"scale(",null,",",null,")");v.push({i:_-4,x:W(c,f)},{i:_-2,x:W(l,d)})}else(f!==1||d!==1)&&m.push(r(m)+"scale("+f+","+d+")")}return function(c,l){var f=[],d=[];return c=t(c),l=t(l),s(c.translateX,c.translateY,l.translateX,l.translateY,f,d),o(c.rotate,l.rotate,f,d),a(c.skewX,l.skewX,f,d),u(c.scaleX,c.scaleY,l.scaleX,l.scaleY,f,d),c=l=null,function(m){for(var v=-1,_=d.length,S;++v<_;)f[(S=d[v]).i]=S.x(m);return f.join("")}}}var br=Qe(xr,"px, ","px)","deg)"),Er=Qe(_r,", ",")",")"),Lr=1e-12;function $e(t){return((t=Math.exp(t))+1/t)/2}function Sr(t){return((t=Math.exp(t))-1/t)/2}function kr(t){return((t=Math.exp(2*t))-1)/(t+1)}const Cr=function t(e,n,i){function r(s,o){var a=s[0],u=s[1],c=s[2],l=o[0],f=o[1],d=o[2],m=l-a,v=f-u,_=m*m+v*v,S,w;if(_<Lr)w=Math.log(d/c)/e,S=function(G){return[a+G*m,u+G*v,c*Math.exp(e*G*w)]};else{var b=Math.sqrt(_),C=(d*d-c*c+i*_)/(2*c*n*b),$=(d*d-c*c-i*_)/(2*d*n*b),z=Math.log(Math.sqrt(C*C+1)-C),N=Math.log(Math.sqrt($*$+1)-$);w=(N-z)/e,S=function(G){var Et=G*w,Lt=$e(z),St=c/(n*b)*(Lt*kr(e*Et+z)-Sr(z));return[a+St*m,u+St*v,c*Lt/$e(e*Et+z)]}}return S.duration=w*1e3*e/Math.SQRT2,S}return r.rho=function(s){var o=Math.max(.001,+s),a=o*o,u=a*a;return t(o,a,u)},r}(Math.SQRT2,2,4);var ut=0,pt=0,ft=0,je=1e3,Ut,mt,qt=0,rt=0,Pt=0,vt=typeof performance=="object"&&performance.now?performance:Date,tn=typeof window=="object"&&window.requestAnimationFrame?window.requestAnimationFrame.bind(window):function(t){setTimeout(t,17)};function de(){return rt||(tn($r),rt=vt.now()+Pt)}function $r(){rt=0}function It(){this._call=this._time=this._next=null}It.prototype=en.prototype={constructor:It,restart:function(t,e,n){if(typeof t!="function")throw new TypeError("callback is not a function");n=(n==null?de():+n)+(e==null?0:+e),!this._next&&mt!==this&&(mt?mt._next=this:Ut=this,mt=this),this._call=t,this._time=n,se()},stop:function(){this._call&&(this._call=null,this._time=1/0,se())}};function en(t,e,n){var i=new It;return i.restart(t,e,n),i}function Tr(){de(),++ut;for(var t=Ut,e;t;)(e=rt-t._time)>=0&&t._call.call(void 0,e),t=t._next;--ut}function Te(){rt=(qt=vt.now())+Pt,ut=pt=0;try{Tr()}finally{ut=0,zr(),rt=0}}function Mr(){var t=vt.now(),e=t-qt;e>je&&(Pt-=e,qt=t)}function zr(){for(var t,e=Ut,n,i=1/0;e;)e._call?(i>e._time&&(i=e._time),t=e,e=e._next):(n=e._next,e._next=null,e=t?t._next=n:Ut=n);mt=t,se(i)}function se(t){if(!ut){pt&&(pt=clearTimeout(pt));var e=t-rt;e>24?(t<1/0&&(pt=setTimeout(Te,t-vt.now()-Pt)),ft&&(ft=clearInterval(ft))):(ft||(qt=vt.now(),ft=setInterval(Mr,je)),ut=1,tn(Te))}}function Me(t,e,n){var i=new It;return e=e==null?0:+e,i.restart(r=>{i.stop(),t(r+e)},e,n),i}var Nr=ce("start","end","cancel","interrupt"),Ar=[],nn=0,ze=1,oe=2,zt=3,Ne=4,ae=5,Nt=6;function Xt(t,e,n,i,r,s){var o=t.__transition;if(!o)t.__transition={};else if(n in o)return;Br(t,n,{name:e,index:i,group:r,on:Nr,tween:Ar,time:s.time,delay:s.delay,duration:s.duration,ease:s.ease,timer:null,state:nn})}function pe(t,e){var n=q(t,e);if(n.state>nn)throw new Error("too late; already scheduled");return n}function V(t,e){var n=q(t,e);if(n.state>zt)throw new Error("too late; already running");return n}function q(t,e){var n=t.__transition;if(!n||!(n=n[e]))throw new Error("transition not found");return n}function Br(t,e,n){var i=t.__transition,r;i[e]=n,n.timer=en(s,0,n.time);function s(c){n.state=ze,n.timer.restart(o,n.delay,n.time),n.delay<=c&&o(c-n.delay)}function o(c){var l,f,d,m;if(n.state!==ze)return u();for(l in i)if(m=i[l],m.name===n.name){if(m.state===zt)return Me(o);m.state===Ne?(m.state=Nt,m.timer.stop(),m.on.call("interrupt",t,t.__data__,m.index,m.group),delete i[l]):+l<e&&(m.state=Nt,m.timer.stop(),m.on.call("cancel",t,t.__data__,m.index,m.group),delete i[l])}if(Me(function(){n.state===zt&&(n.state=Ne,n.timer.restart(a,n.delay,n.time),a(c))}),n.state=oe,n.on.call("start",t,t.__data__,n.index,n.group),n.state===oe){for(n.state=zt,r=new Array(d=n.tween.length),l=0,f=-1;l<d;++l)(m=n.tween[l].value.call(t,t.__data__,n.index,n.group))&&(r[++f]=m);r.length=f+1}}function a(c){for(var l=c<n.duration?n.ease.call(null,c/n.duration):(n.timer.restart(u),n.state=ae,1),f=-1,d=r.length;++f<d;)r[f].call(t,l);n.state===ae&&(n.on.call("end",t,t.__data__,n.index,n.group),u())}function u(){n.state=Nt,n.timer.stop(),delete i[e];for(var c in i)return;delete t.__transition}}function At(t,e){var n=t.__transition,i,r,s=!0,o;if(n){e=e==null?null:e+"";for(o in n){if((i=n[o]).name!==e){s=!1;continue}r=i.state>oe&&i.state<ae,i.state=Nt,i.timer.stop(),i.on.call(r?"interrupt":"cancel",t,t.__data__,i.index,i.group),delete n[o]}s&&delete t.__transition}}function Rr(t){return this.each(function(){At(this,t)})}function Fr(t,e){var n,i;return function(){var r=V(this,t),s=r.tween;if(s!==n){i=n=s;for(var o=0,a=i.length;o<a;++o)if(i[o].name===e){i=i.slice(),i.splice(o,1);break}}r.tween=i}}function Hr(t,e,n){var i,r;if(typeof n!="function")throw new Error;return function(){var s=V(this,t),o=s.tween;if(o!==i){r=(i=o).slice();for(var a={name:e,value:n},u=0,c=r.length;u<c;++u)if(r[u].name===e){r[u]=a;break}u===c&&r.push(a)}s.tween=r}}function Dr(t,e){var n=this._id;if(t+="",arguments.length<2){for(var i=q(this.node(),n).tween,r=0,s=i.length,o;r<s;++r)if((o=i[r]).name===t)return o.value;return null}return this.each((e==null?Fr:Hr)(n,t,e))}function me(t,e,n){var i=t._id;return t.each(function(){var r=V(this,i);(r.value||(r.value={}))[e]=n.apply(this,arguments)}),function(r){return q(r,i).value[e]}}function rn(t,e){var n;return(typeof e=="number"?W:e instanceof wt?ke:(n=wt(e))?(e=n,ke):vr)(t,e)}function Ur(t){return function(){this.removeAttribute(t)}}function qr(t){return function(){this.removeAttributeNS(t.space,t.local)}}function Ir(t,e,n){var i,r=n+"",s;return function(){var o=this.getAttribute(t);return o===r?null:o===i?s:s=e(i=o,n)}}function Or(t,e,n){var i,r=n+"",s;return function(){var o=this.getAttributeNS(t.space,t.local);return o===r?null:o===i?s:s=e(i=o,n)}}function Vr(t,e,n){var i,r,s;return function(){var o,a=n(this),u;return a==null?void this.removeAttribute(t):(o=this.getAttribute(t),u=a+"",o===u?null:o===i&&u===r?s:(r=u,s=e(i=o,a)))}}function Pr(t,e,n){var i,r,s;return function(){var o,a=n(this),u;return a==null?void this.removeAttributeNS(t.space,t.local):(o=this.getAttributeNS(t.space,t.local),u=a+"",o===u?null:o===i&&u===r?s:(r=u,s=e(i=o,a)))}}function Xr(t,e){var n=Vt(t),i=n==="transform"?Er:rn;return this.attrTween(t,typeof e=="function"?(n.local?Pr:Vr)(n,i,me(this,"attr."+t,e)):e==null?(n.local?qr:Ur)(n):(n.local?Or:Ir)(n,i,e))}function Yr(t,e){return function(n){this.setAttribute(t,e.call(this,n))}}function Jr(t,e){return function(n){this.setAttributeNS(t.space,t.local,e.call(this,n))}}function Kr(t,e){var n,i;function r(){var s=e.apply(this,arguments);return s!==i&&(n=(i=s)&&Jr(t,s)),n}return r._value=e,r}function Zr(t,e){var n,i;function r(){var s=e.apply(this,arguments);return s!==i&&(n=(i=s)&&Yr(t,s)),n}return r._value=e,r}function Gr(t,e){var n="attr."+t;if(arguments.length<2)return(n=this.tween(n))&&n._value;if(e==null)return this.tween(n,null);if(typeof e!="function")throw new Error;var i=Vt(t);return this.tween(n,(i.local?Kr:Zr)(i,e))}function Wr(t,e){return function(){pe(this,t).delay=+e.apply(this,arguments)}}function Qr(t,e){return e=+e,function(){pe(this,t).delay=e}}function jr(t){var e=this._id;return arguments.length?this.each((typeof t=="function"?Wr:Qr)(e,t)):q(this.node(),e).delay}function ts(t,e){return function(){V(this,t).duration=+e.apply(this,arguments)}}function es(t,e){return e=+e,function(){V(this,t).duration=e}}function ns(t){var e=this._id;return arguments.length?this.each((typeof t=="function"?ts:es)(e,t)):q(this.node(),e).duration}function is(t,e){if(typeof e!="function")throw new Error;return function(){V(this,t).ease=e}}function rs(t){var e=this._id;return arguments.length?this.each(is(e,t)):q(this.node(),e).ease}function ss(t,e){return function(){var n=e.apply(this,arguments);if(typeof n!="function")throw new Error;V(this,t).ease=n}}function os(t){if(typeof t!="function")throw new Error;return this.each(ss(this._id,t))}function as(t){typeof t!="function"&&(t=He(t));for(var e=this._groups,n=e.length,i=new Array(n),r=0;r<n;++r)for(var s=e[r],o=s.length,a=i[r]=[],u,c=0;c<o;++c)(u=s[c])&&t.call(u,u.__data__,c,s)&&a.push(u);return new Z(i,this._parents,this._name,this._id)}function ls(t){if(t._id!==this._id)throw new Error;for(var e=this._groups,n=t._groups,i=e.length,r=n.length,s=Math.min(i,r),o=new Array(i),a=0;a<s;++a)for(var u=e[a],c=n[a],l=u.length,f=o[a]=new Array(l),d,m=0;m<l;++m)(d=u[m]||c[m])&&(f[m]=d);for(;a<i;++a)o[a]=e[a];return new Z(o,this._parents,this._name,this._id)}function cs(t){return(t+"").trim().split(/^|\s+/).every(function(e){var n=e.indexOf(".");return n>=0&&(e=e.slice(0,n)),!e||e==="start"})}function us(t,e,n){var i,r,s=cs(e)?pe:V;return function(){var o=s(this,t),a=o.on;a!==i&&(r=(i=a).copy()).on(e,n),o.on=r}}function hs(t,e){var n=this._id;return arguments.length<2?q(this.node(),n).on.on(t):this.each(us(n,t,e))}function fs(t){return function(){var e=this.parentNode;for(var n in this.__transition)if(+n!==t)return;e&&e.removeChild(this)}}function ds(){return this.on("end.remove",fs(this._id))}function ps(t){var e=this._name,n=this._id;typeof t!="function"&&(t=ue(t));for(var i=this._groups,r=i.length,s=new Array(r),o=0;o<r;++o)for(var a=i[o],u=a.length,c=s[o]=new Array(u),l,f,d=0;d<u;++d)(l=a[d])&&(f=t.call(l,l.__data__,d,a))&&("__data__"in l&&(f.__data__=l.__data__),c[d]=f,Xt(c[d],e,n,d,c,q(l,n)));return new Z(s,this._parents,e,n)}function ms(t){var e=this._name,n=this._id;typeof t!="function"&&(t=Fe(t));for(var i=this._groups,r=i.length,s=[],o=[],a=0;a<r;++a)for(var u=i[a],c=u.length,l,f=0;f<c;++f)if(l=u[f]){for(var d=t.call(l,l.__data__,f,u),m,v=q(l,n),_=0,S=d.length;_<S;++_)(m=d[_])&&Xt(m,e,n,_,d,v);s.push(d),o.push(l)}return new Z(s,o,e,n)}var gs=_t.prototype.constructor;function ys(){return new gs(this._groups,this._parents)}function ws(t,e){var n,i,r;return function(){var s=ct(this,t),o=(this.style.removeProperty(t),ct(this,t));return s===o?null:s===n&&o===i?r:r=e(n=s,i=o)}}function sn(t){return function(){this.style.removeProperty(t)}}function vs(t,e,n){var i,r=n+"",s;return function(){var o=ct(this,t);return o===r?null:o===i?s:s=e(i=o,n)}}function xs(t,e,n){var i,r,s;return function(){var o=ct(this,t),a=n(this),u=a+"";return a==null&&(u=a=(this.style.removeProperty(t),ct(this,t))),o===u?null:o===i&&u===r?s:(r=u,s=e(i=o,a))}}function _s(t,e){var n,i,r,s="style."+e,o="end."+s,a;return function(){var u=V(this,t),c=u.on,l=u.value[s]==null?a||(a=sn(e)):void 0;(c!==n||r!==l)&&(i=(n=c).copy()).on(o,r=l),u.on=i}}function bs(t,e,n){var i=(t+="")=="transform"?br:rn;return e==null?this.styleTween(t,ws(t,i)).on("end.style."+t,sn(t)):typeof e=="function"?this.styleTween(t,xs(t,i,me(this,"style."+t,e))).each(_s(this._id,t)):this.styleTween(t,vs(t,i,e),n).on("end.style."+t,null)}function Es(t,e,n){return function(i){this.style.setProperty(t,e.call(this,i),n)}}function Ls(t,e,n){var i,r;function s(){var o=e.apply(this,arguments);return o!==r&&(i=(r=o)&&Es(t,o,n)),i}return s._value=e,s}function Ss(t,e,n){var i="style."+(t+="");if(arguments.length<2)return(i=this.tween(i))&&i._value;if(e==null)return this.tween(i,null);if(typeof e!="function")throw new Error;return this.tween(i,Ls(t,e,n??""))}function ks(t){return function(){this.textContent=t}}function Cs(t){return function(){var e=t(this);this.textContent=e??""}}function $s(t){return this.tween("text",typeof t=="function"?Cs(me(this,"text",t)):ks(t==null?"":t+""))}function Ts(t){return function(e){this.textContent=t.call(this,e)}}function Ms(t){var e,n;function i(){var r=t.apply(this,arguments);return r!==n&&(e=(n=r)&&Ts(r)),e}return i._value=t,i}function zs(t){var e="text";if(arguments.length<1)return(e=this.tween(e))&&e._value;if(t==null)return this.tween(e,null);if(typeof t!="function")throw new Error;return this.tween(e,Ms(t))}function Ns(){for(var t=this._name,e=this._id,n=on(),i=this._groups,r=i.length,s=0;s<r;++s)for(var o=i[s],a=o.length,u,c=0;c<a;++c)if(u=o[c]){var l=q(u,e);Xt(u,t,n,c,o,{time:l.time+l.delay+l.duration,delay:0,duration:l.duration,ease:l.ease})}return new Z(i,this._parents,t,n)}function As(){var t,e,n=this,i=n._id,r=n.size();return new Promise(function(s,o){var a={value:o},u={value:function(){--r===0&&s()}};n.each(function(){var c=V(this,i),l=c.on;l!==t&&(e=(t=l).copy(),e._.cancel.push(a),e._.interrupt.push(a),e._.end.push(u)),c.on=e}),r===0&&s()})}var Bs=0;function Z(t,e,n,i){this._groups=t,this._parents=e,this._name=n,this._id=i}function on(){return++Bs}var Y=_t.prototype;Z.prototype={constructor:Z,select:ps,selectAll:ms,selectChild:Y.selectChild,selectChildren:Y.selectChildren,filter:as,merge:ls,selection:ys,transition:Ns,call:Y.call,nodes:Y.nodes,node:Y.node,size:Y.size,empty:Y.empty,each:Y.each,on:hs,attr:Xr,attrTween:Gr,style:bs,styleTween:Ss,text:$s,textTween:zs,remove:ds,tween:Dr,delay:jr,duration:ns,ease:rs,easeVarying:os,end:As,[Symbol.iterator]:Y[Symbol.iterator]};function Rs(t){return((t*=2)<=1?t*t*t:(t-=2)*t*t+2)/2}var Fs={time:null,delay:0,duration:250,ease:Rs};function Hs(t,e){for(var n;!(n=t.__transition)||!(n=n[e]);)if(!(t=t.parentNode))throw new Error(`transition ${e} not found`);return n}function Ds(t){var e,n;t instanceof Z?(e=t._id,t=t._name):(e=on(),(n=Fs).time=de(),t=t==null?null:t+"");for(var i=this._groups,r=i.length,s=0;s<r;++s)for(var o=i[s],a=o.length,u,c=0;c<a;++c)(u=o[c])&&Xt(u,t,e,c,o,n||Hs(u,e));return new Z(i,this._parents,t,e)}_t.prototype.interrupt=Rr;_t.prototype.transition=Ds;function Us(t){var e=0,n=t.children,i=n&&n.length;if(!i)e=1;else for(;--i>=0;)e+=n[i].value;t.value=e}function qs(){return this.eachAfter(Us)}function Is(t,e){let n=-1;for(const i of this)t.call(e,i,++n,this);return this}function Os(t,e){for(var n=this,i=[n],r,s,o=-1;n=i.pop();)if(t.call(e,n,++o,this),r=n.children)for(s=r.length-1;s>=0;--s)i.push(r[s]);return this}function Vs(t,e){for(var n=this,i=[n],r=[],s,o,a,u=-1;n=i.pop();)if(r.push(n),s=n.children)for(o=0,a=s.length;o<a;++o)i.push(s[o]);for(;n=r.pop();)t.call(e,n,++u,this);return this}function Ps(t,e){let n=-1;for(const i of this)if(t.call(e,i,++n,this))return i}function Xs(t){return this.eachAfter(function(e){for(var n=+t(e.data)||0,i=e.children,r=i&&i.length;--r>=0;)n+=i[r].value;e.value=n})}function Ys(t){return this.eachBefore(function(e){e.children&&e.children.sort(t)})}function Js(t){for(var e=this,n=Ks(e,t),i=[e];e!==n;)e=e.parent,i.push(e);for(var r=i.length;t!==n;)i.splice(r,0,t),t=t.parent;return i}function Ks(t,e){if(t===e)return t;var n=t.ancestors(),i=e.ancestors(),r=null;for(t=n.pop(),e=i.pop();t===e;)r=t,t=n.pop(),e=i.pop();return r}function Zs(){for(var t=this,e=[t];t=t.parent;)e.push(t);return e}function Gs(){return Array.from(this)}function Ws(){var t=[];return this.eachBefore(function(e){e.children||t.push(e)}),t}function Qs(){var t=this,e=[];return t.each(function(n){n!==t&&e.push({source:n.parent,target:n})}),e}function*js(){var t=this,e,n=[t],i,r,s;do for(e=n.reverse(),n=[];t=e.pop();)if(yield t,i=t.children)for(r=0,s=i.length;r<s;++r)n.push(i[r]);while(n.length)}function st(t,e){t instanceof Map?(t=[void 0,t],e===void 0&&(e=no)):e===void 0&&(e=eo);for(var n=new xt(t),i,r=[n],s,o,a,u;i=r.pop();)if((o=e(i.data))&&(u=(o=Array.from(o)).length))for(i.children=o,a=u-1;a>=0;--a)r.push(s=o[a]=new xt(o[a])),s.parent=i,s.depth=i.depth+1;return n.eachBefore(ro)}function to(){return st(this).eachBefore(io)}function eo(t){return t.children}function no(t){return Array.isArray(t)?t[1]:null}function io(t){t.data.value!==void 0&&(t.value=t.data.value),t.data=t.data.data}function ro(t){var e=0;do t.height=e;while((t=t.parent)&&t.height<++e)}function xt(t){this.data=t,this.depth=this.height=0,this.parent=null}xt.prototype=st.prototype={constructor:xt,count:qs,each:Is,eachAfter:Vs,eachBefore:Os,find:Ps,sum:Xs,sort:Ys,path:Js,ancestors:Zs,descendants:Gs,leaves:Ws,links:Qs,copy:to,[Symbol.iterator]:js};function so(t,e){return t.parent===e.parent?1:2}function Gt(t){var e=t.children;return e?e[0]:t.t}function Wt(t){var e=t.children;return e?e[e.length-1]:t.t}function oo(t,e,n){var i=n/(e.i-t.i);e.c-=i,e.s+=n,t.c+=i,e.z+=n,e.m+=n}function ao(t){for(var e=0,n=0,i=t.children,r=i.length,s;--r>=0;)s=i[r],s.z+=e,s.m+=e,e+=s.s+(n+=s.c)}function lo(t,e,n){return t.a.parent===e.parent?t.a:n}function Bt(t,e){this._=t,this.parent=null,this.children=null,this.A=null,this.a=this,this.z=0,this.m=0,this.c=0,this.s=0,this.t=null,this.i=e}Bt.prototype=Object.create(xt.prototype);function co(t){for(var e=new Bt(t,0),n,i=[e],r,s,o,a;n=i.pop();)if(s=n._.children)for(n.children=new Array(a=s.length),o=a-1;o>=0;--o)i.push(r=n.children[o]=new Bt(s[o],o)),r.parent=n;return(e.parent=new Bt(null,0)).children=[e],e}function Ae(){var t=so,e=1,n=1,i=null;function r(c){var l=co(c);if(l.eachAfter(s),l.parent.m=-l.z,l.eachBefore(o),i)c.eachBefore(u);else{var f=c,d=c,m=c;c.eachBefore(function(b){b.x<f.x&&(f=b),b.x>d.x&&(d=b),b.depth>m.depth&&(m=b)});var v=f===d?1:t(f,d)/2,_=v-f.x,S=e/(d.x+v+_),w=n/(m.depth||1);c.eachBefore(function(b){b.x=(b.x+_)*S,b.y=b.depth*w})}return c}function s(c){var l=c.children,f=c.parent.children,d=c.i?f[c.i-1]:null;if(l){ao(c);var m=(l[0].z+l[l.length-1].z)/2;d?(c.z=d.z+t(c._,d._),c.m=c.z-m):c.z=m}else d&&(c.z=d.z+t(c._,d._));c.parent.A=a(c,d,c.parent.A||f[0])}function o(c){c._.x=c.z+c.parent.m,c.m+=c.parent.m}function a(c,l,f){if(l){for(var d=c,m=c,v=l,_=d.parent.children[0],S=d.m,w=m.m,b=v.m,C=_.m,$;v=Wt(v),d=Gt(d),v&&d;)_=Gt(_),m=Wt(m),m.a=c,$=v.z+b-d.z-S+t(v._,d._),$>0&&(oo(lo(v,c,f),c,$),S+=$,w+=$),b+=v.m,S+=d.m,C+=_.m,w+=m.m;v&&!Wt(m)&&(m.t=v,m.m+=b-w),d&&!Gt(_)&&(_.t=d,_.m+=S-C,f=c)}return f}function u(c){c.x*=e,c.y=c.depth*n}return r.separation=function(c){return arguments.length?(t=c,r):t},r.size=function(c){return arguments.length?(i=!1,e=+c[0],n=+c[1],r):i?null:[e,n]},r.nodeSize=function(c){return arguments.length?(i=!0,e=+c[0],n=+c[1],r):i?[e,n]:null},r}const Tt=t=>()=>t;function uo(t,{sourceEvent:e,target:n,transform:i,dispatch:r}){Object.defineProperties(this,{type:{value:t,enumerable:!0,configurable:!0},sourceEvent:{value:e,enumerable:!0,configurable:!0},target:{value:n,enumerable:!0,configurable:!0},transform:{value:i,enumerable:!0,configurable:!0},_:{value:r}})}function J(t,e,n){this.k=t,this.x=e,this.y=n}J.prototype={constructor:J,scale:function(t){return t===1?this:new J(this.k*t,this.x,this.y)},translate:function(t,e){return t===0&e===0?this:new J(this.k,this.x+this.k*t,this.y+this.k*e)},apply:function(t){return[t[0]*this.k+this.x,t[1]*this.k+this.y]},applyX:function(t){return t*this.k+this.x},applyY:function(t){return t*this.k+this.y},invert:function(t){return[(t[0]-this.x)/this.k,(t[1]-this.y)/this.k]},invertX:function(t){return(t-this.x)/this.k},invertY:function(t){return(t-this.y)/this.k},rescaleX:function(t){return t.copy().domain(t.range().map(this.invertX,this).map(t.invert,t))},rescaleY:function(t){return t.copy().domain(t.range().map(this.invertY,this).map(t.invert,t))},toString:function(){return"translate("+this.x+","+this.y+") scale("+this.k+")"}};var lt=new J(1,0,0);le.prototype=J.prototype;function le(t){for(;!t.__zoom;)if(!(t=t.parentNode))return lt;return t.__zoom}function Qt(t){t.stopImmediatePropagation()}function dt(t){t.preventDefault(),t.stopImmediatePropagation()}function ho(t){return(!t.ctrlKey||t.type==="wheel")&&!t.button}function fo(){var t=this;return t instanceof SVGElement?(t=t.ownerSVGElement||t,t.hasAttribute("viewBox")?(t=t.viewBox.baseVal,[[t.x,t.y],[t.x+t.width,t.y+t.height]]):[[0,0],[t.width.baseVal.value,t.height.baseVal.value]]):[[0,0],[t.clientWidth,t.clientHeight]]}function Be(){return this.__zoom||lt}function po(t){return-t.deltaY*(t.deltaMode===1?.05:t.deltaMode?1:.002)*(t.ctrlKey?10:1)}function mo(){return navigator.maxTouchPoints||"ontouchstart"in this}function go(t,e,n){var i=t.invertX(e[0][0])-n[0][0],r=t.invertX(e[1][0])-n[1][0],s=t.invertY(e[0][1])-n[0][1],o=t.invertY(e[1][1])-n[1][1];return t.translate(r>i?(i+r)/2:Math.min(0,i)||Math.max(0,r),o>s?(s+o)/2:Math.min(0,s)||Math.max(0,o))}function yo(){var t=ho,e=fo,n=go,i=po,r=mo,s=[0,1/0],o=[[-1/0,-1/0],[1/0,1/0]],a=250,u=Cr,c=ce("start","zoom","end"),l,f,d,m=500,v=150,_=0,S=10;function w(h){h.property("__zoom",Be).on("wheel.zoom",Et,{passive:!1}).on("mousedown.zoom",Lt).on("dblclick.zoom",St).filter(r).on("touchstart.zoom",an).on("touchmove.zoom",ln).on("touchend.zoom touchcancel.zoom",cn).style("-webkit-tap-highlight-color","rgba(0,0,0,0)")}w.transform=function(h,g,p,y){var x=h.selection?h.selection():h;x.property("__zoom",Be),h!==x?z(h,g,p,y):x.interrupt().each(function(){N(this,arguments).event(y).start().zoom(null,typeof g=="function"?g.apply(this,arguments):g).end()})},w.scaleBy=function(h,g,p,y){w.scaleTo(h,function(){var x=this.__zoom.k,E=typeof g=="function"?g.apply(this,arguments):g;return x*E},p,y)},w.scaleTo=function(h,g,p,y){w.transform(h,function(){var x=e.apply(this,arguments),E=this.__zoom,L=p==null?$(x):typeof p=="function"?p.apply(this,arguments):p,k=E.invert(L),T=typeof g=="function"?g.apply(this,arguments):g;return n(C(b(E,T),L,k),x,o)},p,y)},w.translateBy=function(h,g,p,y){w.transform(h,function(){return n(this.__zoom.translate(typeof g=="function"?g.apply(this,arguments):g,typeof p=="function"?p.apply(this,arguments):p),e.apply(this,arguments),o)},null,y)},w.translateTo=function(h,g,p,y,x){w.transform(h,function(){var E=e.apply(this,arguments),L=this.__zoom,k=y==null?$(E):typeof y=="function"?y.apply(this,arguments):y;return n(lt.translate(k[0],k[1]).scale(L.k).translate(typeof g=="function"?-g.apply(this,arguments):-g,typeof p=="function"?-p.apply(this,arguments):-p),E,o)},y,x)};function b(h,g){return g=Math.max(s[0],Math.min(s[1],g)),g===h.k?h:new J(g,h.x,h.y)}function C(h,g,p){var y=g[0]-p[0]*h.k,x=g[1]-p[1]*h.k;return y===h.x&&x===h.y?h:new J(h.k,y,x)}function $(h){return[(+h[0][0]+ +h[1][0])/2,(+h[0][1]+ +h[1][1])/2]}function z(h,g,p,y){h.on("start.zoom",function(){N(this,arguments).event(y).start()}).on("interrupt.zoom end.zoom",function(){N(this,arguments).event(y).end()}).tween("zoom",function(){var x=this,E=arguments,L=N(x,E).event(y),k=e.apply(x,E),T=p==null?$(k):typeof p=="function"?p.apply(x,E):p,I=Math.max(k[1][0]-k[0][0],k[1][1]-k[0][1]),M=x.__zoom,F=typeof g=="function"?g.apply(x,E):g,P=u(M.invert(T).concat(I/M.k),F.invert(T).concat(I/F.k));return function(H){if(H===1)H=F;else{var X=P(H),Yt=I/X[2];H=new J(Yt,T[0]-X[0]*Yt,T[1]-X[1]*Yt)}L.zoom(null,H)}})}function N(h,g,p){return!p&&h.__zooming||new G(h,g)}function G(h,g){this.that=h,this.args=g,this.active=0,this.sourceEvent=null,this.extent=e.apply(h,g),this.taps=0}G.prototype={event:function(h){return h&&(this.sourceEvent=h),this},start:function(){return++this.active===1&&(this.that.__zooming=this,this.emit("start")),this},zoom:function(h,g){return this.mouse&&h!=="mouse"&&(this.mouse[1]=g.invert(this.mouse[0])),this.touch0&&h!=="touch"&&(this.touch0[1]=g.invert(this.touch0[0])),this.touch1&&h!=="touch"&&(this.touch1[1]=g.invert(this.touch1[0])),this.that.__zoom=g,this.emit("zoom"),this},end:function(){return--this.active===0&&(delete this.that.__zooming,this.emit("end")),this},emit:function(h){var g=Q(this.that).datum();c.call(h,this.that,new uo(h,{sourceEvent:this.sourceEvent,target:w,transform:this.that.__zoom,dispatch:c}),g)}};function Et(h,...g){if(!t.apply(this,arguments))return;var p=N(this,g).event(h),y=this.__zoom,x=Math.max(s[0],Math.min(s[1],y.k*Math.pow(2,i.apply(this,arguments)))),E=j(h);if(p.wheel)(p.mouse[0][0]!==E[0]||p.mouse[0][1]!==E[1])&&(p.mouse[1]=y.invert(p.mouse[0]=E)),clearTimeout(p.wheel);else{if(y.k===x)return;p.mouse=[E,y.invert(E)],At(this),p.start()}dt(h),p.wheel=setTimeout(L,v),p.zoom("mouse",n(C(b(y,x),p.mouse[0],p.mouse[1]),p.extent,o));function L(){p.wheel=null,p.end()}}function Lt(h,...g){if(d||!t.apply(this,arguments))return;var p=h.currentTarget,y=N(this,g,!0).event(h),x=Q(h.view).on("mousemove.zoom",T,!0).on("mouseup.zoom",I,!0),E=j(h,p),L=h.clientX,k=h.clientY;tr(h.view),Qt(h),y.mouse=[E,this.__zoom.invert(E)],At(this),y.start();function T(M){if(dt(M),!y.moved){var F=M.clientX-L,P=M.clientY-k;y.moved=F*F+P*P>_}y.event(M).zoom("mouse",n(C(y.that.__zoom,y.mouse[0]=j(M,p),y.mouse[1]),y.extent,o))}function I(M){x.on("mousemove.zoom mouseup.zoom",null),er(M.view,y.moved),dt(M),y.event(M).end()}}function St(h,...g){if(t.apply(this,arguments)){var p=this.__zoom,y=j(h.changedTouches?h.changedTouches[0]:h,this),x=p.invert(y),E=p.k*(h.shiftKey?.5:2),L=n(C(b(p,E),y,x),e.apply(this,g),o);dt(h),a>0?Q(this).transition().duration(a).call(z,L,y,h):Q(this).call(w.transform,L,y,h)}}function an(h,...g){if(t.apply(this,arguments)){var p=h.touches,y=p.length,x=N(this,g,h.changedTouches.length===y).event(h),E,L,k,T;for(Qt(h),L=0;L<y;++L)k=p[L],T=j(k,this),T=[T,this.__zoom.invert(T),k.identifier],x.touch0?!x.touch1&&x.touch0[2]!==T[2]&&(x.touch1=T,x.taps=0):(x.touch0=T,E=!0,x.taps=1+!!l);l&&(l=clearTimeout(l)),E&&(x.taps<2&&(f=T[0],l=setTimeout(function(){l=null},m)),At(this),x.start())}}function ln(h,...g){if(this.__zooming){var p=N(this,g).event(h),y=h.changedTouches,x=y.length,E,L,k,T;for(dt(h),E=0;E<x;++E)L=y[E],k=j(L,this),p.touch0&&p.touch0[2]===L.identifier?p.touch0[0]=k:p.touch1&&p.touch1[2]===L.identifier&&(p.touch1[0]=k);if(L=p.that.__zoom,p.touch1){var I=p.touch0[0],M=p.touch0[1],F=p.touch1[0],P=p.touch1[1],H=(H=F[0]-I[0])*H+(H=F[1]-I[1])*H,X=(X=P[0]-M[0])*X+(X=P[1]-M[1])*X;L=b(L,Math.sqrt(H/X)),k=[(I[0]+F[0])/2,(I[1]+F[1])/2],T=[(M[0]+P[0])/2,(M[1]+P[1])/2]}else if(p.touch0)k=p.touch0[0],T=p.touch0[1];else return;p.zoom("touch",n(C(L,k,T),p.extent,o))}}function cn(h,...g){if(this.__zooming){var p=N(this,g).event(h),y=h.changedTouches,x=y.length,E,L;for(Qt(h),d&&clearTimeout(d),d=setTimeout(function(){d=null},m),E=0;E<x;++E)L=y[E],p.touch0&&p.touch0[2]===L.identifier?delete p.touch0:p.touch1&&p.touch1[2]===L.identifier&&delete p.touch1;if(p.touch1&&!p.touch0&&(p.touch0=p.touch1,delete p.touch1),p.touch0)p.touch0[1]=this.__zoom.invert(p.touch0[0]);else if(p.end(),p.taps===2&&(L=j(L,this),Math.hypot(f[0]-L[0],f[1]-L[1])<S)){var k=Q(this).on("dblclick.zoom");k&&k.apply(this,arguments)}}}return w.wheelDelta=function(h){return arguments.length?(i=typeof h=="function"?h:Tt(+h),w):i},w.filter=function(h){return arguments.length?(t=typeof h=="function"?h:Tt(!!h),w):t},w.touchable=function(h){return arguments.length?(r=typeof h=="function"?h:Tt(!!h),w):r},w.extent=function(h){return arguments.length?(e=typeof h=="function"?h:Tt([[+h[0][0],+h[0][1]],[+h[1][0],+h[1][1]]]),w):e},w.scaleExtent=function(h){return arguments.length?(s[0]=+h[0],s[1]=+h[1],w):[s[0],s[1]]},w.translateExtent=function(h){return arguments.length?(o[0][0]=+h[0][0],o[1][0]=+h[1][0],o[0][1]=+h[0][1],o[1][1]=+h[1][1],w):[[o[0][0],o[0][1]],[o[1][0],o[1][1]]]},w.constrain=function(h){return arguments.length?(n=h,w):n},w.duration=function(h){return arguments.length?(a=+h,w):a},w.interpolate=function(h){return arguments.length?(u=h,w):u},w.on=function(){var h=c.on.apply(c,arguments);return h===c?w:h},w.clickDistance=function(h){return arguments.length?(_=(h=+h)*h,w):Math.sqrt(_)},w.tapDistance=function(h){return arguments.length?(S=+h,w):S},w}function wo(t,e){let n;return(...i)=>{n||(t(...i),n=!0,setTimeout(()=>n=!1,e))}}class vo{constructor(e){this.container=e,this.svg=null,this.g=null,this.width=e.clientWidth||1e3,this.height=Math.max(window.innerHeight-200,600),this.expandedNodes=new Set,this.zoomLevel=1,this.nodeId=0,this.zoomBehavior=null,this.treeData=null,this.listeners=new Map,this.renderControls()}renderControls(){const e=document.createElement("div");e.className="diagram-controls",e.innerHTML=`
-      <div class="control-group">
-        <button id="zoomInBtn" title="Zoom In" aria-label="Zoom in">
-          <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2">
-            <path d="M12 5v14M5 12h14"/>
-          </svg>
-        </button>
-        <button id="zoomOutBtn" title="Zoom Out" aria-label="Zoom out">
-          <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2">
-            <path d="M5 12h14"/>
-          </svg>
-        </button>
-        <span id="zoomLevel" class="zoom-level" aria-live="polite">100%</span>
-      </div>
-      <div class="control-group">
-        <button id="expandAllBtn" title="Expand All" aria-label="Expand all nodes">
-          <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2">
-            <path d="M4 4h16v16H4zM8 12h8M12 8v8"/>
-          </svg>
-        </button>
-        <button id="collapseAllBtn" title="Collapse All" aria-label="Collapse all nodes">
-          <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2">
-            <path d="M4 4h16v16H4zM8 12h8"/>
-          </svg>
-        </button>
-      </div>
-      <div class="control-group">
-        <button id="resetViewBtn" title="Reset View" aria-label="Reset view">
-          <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2">
-            <path d="M4 4c5-5 14-5 19 0M4 20c5 5 14 5 19 0"/>
-          </svg>
-        </button>
-      </div>
-    `,this.container.appendChild(e),this.setupControlListeners()}setupControlListeners(){this.cleanupListeners("controls"),Object.entries({"#zoomInBtn":()=>this.zoom(1.2),"#zoomOutBtn":()=>this.zoom(.8),"#expandAllBtn":()=>this.expandAll(),"#collapseAllBtn":()=>this.collapseAll(),"#resetViewBtn":()=>this.resetView()}).forEach(([n,i])=>{const r=this.container.querySelector(n);if(r){const s=()=>i();r.addEventListener("click",s),this.listeners.set(r,[["click",s]])}})}cleanupListeners(e="all"){this.listeners.forEach((n,i)=>{(e==="all"||e==="controls"&&i.tagName==="BUTTON")&&(n.forEach(([r,s])=>i.removeEventListener(r,s)),this.listeners.delete(i))})}render(e,n={},i="url",r=""){U(50),this.container.innerHTML="",this.svg=null,this.g=null,this.treeData=null,this.renderControls();const s=this.filterUrls(e,n,r),o=this.sortUrls(s,i);this.treeData=this.buildTree(o);try{this.drawDiagram()}catch(a){console.error("Failed to render diagram:",a)}K()}filterUrls(e,n,i){return e.filter(r=>{const s=parseFloat(r.priority)||0,o=r.lastmod?new Date(r.lastmod):null,a=parseFloat(n.priority)||0,u=n.lastmod?new Date(n.lastmod):null;return(!i||r.loc.toLowerCase().includes(i.toLowerCase()))&&s>=a&&(!u||o&&o>=u)})}sortUrls(e,n){return e.sort((i,r)=>n==="lastmod"?new Date(r.lastmod||0)-new Date(i.lastmod||0)||i.loc.localeCompare(r.loc):n==="priority"&&parseFloat(r.priority||0)-parseFloat(i.priority||0)||i.loc.localeCompare(r.loc))}buildTree(e){const n={name:"Sitemap",children:[],isLazy:!0},i=new Map;return e.forEach(r=>{try{const o=new URL(r.loc||"").hostname;let a=i.get(o);a||(a={name:o,children:[],isLazy:!0},i.set(o,a),n.children.push(a)),a.children.length||a.children.push({name:"URLs",urlCount:0,children:[],isLazy:!0}),a.children[0].urlCount=(a.children[0].urlCount||0)+1,a.children[0].urls=a.children[0].urls||[],a.children[0].urls.push(r)}catch(s){console.warn(`Skipping invalid URL: ${r.loc}`,s)}}),n}async loadLazyChildren(e){e.data.isLazy&&e.data.urls&&!e.children&&(e.children=e.data.urls.map(n=>({name:n.loc.split("/").pop(),url:n,children:[]})),e.data.isLazy=!1)}drawDiagram(){const e={top:60,right:120,bottom:60,left:120};this.width=this.container.clientWidth-e.left-e.right,this.height=this.container.clientHeight-e.top-e.bottom,this.svg=Q(this.container).append("svg").attr("width",this.width+e.left+e.right).attr("height",this.height+e.top+e.bottom).attr("role","tree").style("background","#fafafa"),this.g=this.svg.append("g").attr("transform",`translate(${e.left},${e.top})`),this.zoomBehavior=yo().scaleExtent([.3,5]).on("zoom",wo(s=>this.zoomed(s),50)),this.svg.call(this.zoomBehavior).on("dblclick.zoom",null).on("wheel.zoom",null),this.svg.on("wheel",s=>{s.preventDefault();const o=s.deltaY<0?1.1:.9;this.zoom(o,s.clientX,s.clientY)});const n=st(this.treeData),r=Ae().nodeSize([60,250])(n);r.x0=this.width/2,r.y0=0,r.children.forEach(s=>this.collapse(s,!0)),this.g.datum(r),this.update(r),this.fitToView(r)}zoomed(e){this.g&&(this.g.attr("transform",e.transform),this.zoomLevel=e.transform.k,this.updateZoomLevelDisplay())}zoom(e,n=null,i=null){if(!this.svg||!this.zoomBehavior)return;const r=le(this.svg.node());let s=r.k*e;if(n&&i){const o=this.svg.node().getBoundingClientRect(),a=n-o.left,u=i-o.top,c=a-(a-r.x)*e/r.k,l=u-(u-r.y)*e/r.k;this.svg.transition().duration(300).call(this.zoomBehavior.transform,lt.translate(c,l).scale(s))}else this.svg.transition().duration(300).call(this.zoomBehavior.scaleBy,e)}updateZoomLevelDisplay(){const e=Math.round(this.zoomLevel*100),n=this.container.querySelector("#zoomLevel");n&&(n.textContent=`${e}%`)}update(e){const r=Ae().nodeSize([60,250])(e),s=this.g.selectAll(".link").data(r.links(),l=>l.target.id),o=s.enter().append("path").attr("class","link").attr("d",l=>this.diagonal({x:e.x0,y:e.y0},{x:e.x0,y:e.y0}));s.merge(o).transition().duration(650).attr("d",l=>this.diagonal(l.source,l.target)),s.exit().transition().duration(650).attr("d",l=>this.diagonal({x:e.x,y:e.y},{x:e.x,y:e.y})).remove();const a=this.g.selectAll(".node").data(r.descendants(),l=>l.id||(l.id=++this.nodeId)),u=a.enter().append("g").attr("class","node").attr("transform",l=>`translate(${e.y0},${e.x0})`).on("click",async(l,f)=>{f.data.isLazy&&!f.children&&!f._children?(U(50),await this.loadLazyChildren(f),K(),f.children&&this.update(f)):f.children?(f._children=f.children,f.children=null,this.expandedNodes.delete(f.data.name)):f._children&&(f.children=f._children,f._children=null,this.expandedNodes.add(f.data.name)),this.update(f),this.centerNode(f)});u.filter(l=>l.children||l._children||l.data.isLazy).append("rect").attr("class","folder-bg").attr("width",160).attr("height",40).attr("x",-80).attr("y",-20).attr("rx",8),u.filter(l=>l.children||l._children||l.data.isLazy).append("path").attr("class","folder-icon").attr("d","M10 4H4c-1.1 0-2 .9-2 2v12c0 1.1.9 2 2 2h16c1.1 0 2-.9 2-2V8c0-1.1-.9-2-2-2h-8L10 4z").attr("transform","translate(-65, -18) scale(1.2)"),u.append("text").attr("dy","0.35em").attr("x",l=>l.children||l._children||l.data.isLazy?-30:10).attr("text-anchor",l=>l.children||l._children||l.data.isLazy?"end":"start").attr("class",l=>l.children||l._children||l.data.isLazy?"folder-label":"url-label").text(l=>l.data.urlCount?`${l.data.name} (${l.data.urlCount})`:l.data.name.length>25?`${l.data.name.slice(0,22)}...`:l.data.name).on("click",(l,f)=>{l.stopPropagation(),f.data.url&&window.open(f.data.url.loc,"_blank")}),u.append("title").text(l=>l.data.url?this.formatUrl(l.data.url):l.data.name);const c=a.merge(u).transition().duration(650).attr("transform",l=>`translate(${l.y},${l.x})`);c.select(".folder-bg").attr("x",-80).attr("y",-20),c.select(".folder-icon").attr("transform","translate(-65, -18) scale(1.2)"),c.select("text").attr("x",l=>l.children||l._children||l.data.isLazy?-30:10).attr("text-anchor",l=>l.children||l._children||l.data.isLazy?"end":"start"),a.exit().transition().duration(650).attr("transform",l=>`translate(${e.y},${e.x})`).remove(),r.each(l=>{l.x0=l.x,l.y0=l.y})}diagonal(e,n){return`M ${e.y} ${e.x} C ${e.y} ${(e.x+n.x)/2}, ${n.y} ${(e.x+n.x)/2}, ${n.y} ${n.x}`}formatUrl(e){let n=e.loc;return e.lastmod&&(n+=`
-Last Modified: ${e.lastmod}`),e.changefreq&&(n+=`
-Change Frequency: ${e.changefreq}`),e.priority&&(n+=`
-Priority: ${e.priority}`),n}centerNode(e){if(!this.svg||!this.zoomBehavior||!this.g)return;const i=le(this.svg.node()).k,r=-e.y*i+this.width/2,s=-e.x*i+this.height/2;this.svg.transition().duration(650).call(this.zoomBehavior.transform,lt.translate(r,s).scale(i))}expandAll(){if(!this.g)return;const e=st(this.g.datum());e.descendants().forEach(async n=>{n.data.isLazy&&await this.loadLazyChildren(n),n._children&&(n.children=n._children,n._children=null,this.expandedNodes.add(n.data.name))}),this.update(e),this.fitToView(e)}collapseAll(){if(!this.g)return;const e=st(this.g.datum());e.descendants().forEach(n=>this.collapse(n)),this.update(e),this.fitToView(e)}collapse(e,n=!1){e.children&&(e._children=e.children,e._children.forEach(i=>this.collapse(i)),e.children=null,n||this.expandedNodes.delete(e.data.name))}fitToView(e){if(!this.zoomBehavior||!this.g)return;const n=this.getTreeBounds(e),i=n.maxY-n.minY,r=n.maxX-n.minX,s=this.width/(i||1),o=this.height/(r||1),a=Math.min(s,o,2)*.9,u=this.width/2-(n.minY+i/2)*a,c=this.height/2-(n.minX+r/2)*a;this.svg.transition().duration(650).call(this.zoomBehavior.transform,lt.translate(u,c).scale(a)),this.zoomLevel=a,this.updateZoomLevelDisplay()}getTreeBounds(e){let n=1/0,i=-1/0,r=1/0,s=-1/0;return e.each(o=>{n=Math.min(n,o.x),i=Math.max(i,o.x),r=Math.min(r,o.y),s=Math.max(s,o.y)}),{minX:n,maxX:i,minY:r,maxY:s}}resetView(){if(!this.g)return;const e=st(this.g.datum());this.fitToView(e)}destroy(){this.svg&&(this.svg.on(".zoom",null),this.svg.on("wheel",null)),this.cleanupListeners(),this.container.innerHTML="",this.svg=null,this.g=null,this.zoomBehavior=null,this.treeData=null,this.expandedNodes.clear()}}function xo(t,e){let n;return(...i)=>{clearTimeout(n),n=setTimeout(()=>t(...i),e)}}class _o{constructor(e,n={}){this.container=e,this.options={onApply:n.onApply||(()=>{}),onReset:n.onReset||(()=>{}),onSearch:n.onSearch||(()=>{}),...n},this.isExpanded=!1,this.state={search:"",priority:"0",lastmod:"",sortBy:"url"},this.listeners=new Map,this.render(),this.setupEventListeners()}render(){this.container.innerHTML=`
-      <div class="search-container">
-        <input type="text" id="urlSearch" placeholder="Search URLs..." value="${this.state.search}" aria-label="Search URLs">
-        <span id="toggleFiltersLink" class="filter-toggle" aria-expanded="${this.isExpanded}" role="button" tabindex="0">
-          ${this.isExpanded?"Hide filters":"Show filters"}
-        </span>
-      </div>
-      <div id="additional-filters" class="${this.isExpanded?"":"hidden"}">
-        <div class="filter-group">
-          <label for="priorityFilter">Min Priority:</label>
-          <input type="number" id="priorityFilter" min="0" max="1" step="0.1" value="${this.state.priority}" aria-label="Minimum priority filter">
-        </div>
-        <div class="filter-group">
-          <label for="lastmodFilter">Last Modified After:</label>
-          <input type="date" id="lastmodFilter" value="${this.state.lastmod}" aria-label="Last modified after date filter">
-        </div>
-        <div class="filter-group">
-          <label for="sortBy">Sort By:</label>
-          <select id="sortBy" aria-label="Sort by">
-            <option value="url" ${this.state.sortBy==="url"?"selected":""}>URL</option>
-            <option value="lastmod" ${this.state.sortBy==="lastmod"?"selected":""}>Last Modified</option>
-            <option value="priority" ${this.state.sortBy==="priority"?"selected":""}>Priority</option>
-          </select>
-        </div>
-        <div class="filter-actions">
-          <button id="applyFiltersBtn" aria-label="Apply filters">Apply</button>
-          <button id="resetFiltersBtn" aria-label="Reset filters">Reset</button>
-        </div>
-      </div>
-    `}setupEventListeners(){this.cleanupListeners();const e=this.container.querySelector("#urlSearch");if(e){const s=xo(()=>{this.state.search=e.value.toLowerCase(),this.options.onSearch(this.state.search)},300);e.addEventListener("input",s),this.listeners.set(e,[["input",s]])}const n=this.container.querySelector("#toggleFiltersLink");if(n){const s=()=>{this.isExpanded=!this.isExpanded,this.render(),this.setupEventListeners(),n.focus()};n.addEventListener("click",s),n.addEventListener("keydown",o=>o.key==="Enter"&&s()),this.listeners.set(n,[["click",s],["keydown",o=>o.key==="Enter"&&s()]])}const i=this.container.querySelector("#applyFiltersBtn");if(i){const s=()=>{const o=Math.min(1,Math.max(0,parseFloat(this.container.querySelector("#priorityFilter").value)||0)).toString(),a=this.container.querySelector("#lastmodFilter").value,u=this.container.querySelector("#sortBy").value;(o!==this.state.priority||a!==this.state.lastmod||u!==this.state.sortBy)&&(this.state.priority=o,this.state.lastmod=a,this.state.sortBy=u,this.options.onApply({priority:this.state.priority,lastmod:this.state.lastmod},this.state.sortBy))};i.addEventListener("click",s),this.listeners.set(i,[["click",s]])}const r=this.container.querySelector("#resetFiltersBtn");if(r){const s=()=>{this.state={search:"",priority:"0",lastmod:"",sortBy:"url"},this.render(),this.setupEventListeners(),this.options.onReset()};r.addEventListener("click",s),this.listeners.set(r,[["click",s]])}}cleanupListeners(){this.listeners.forEach((e,n)=>{e.forEach(([i,r])=>n.removeEventListener(i,r))}),this.listeners.clear()}toggleVisibility(e){this.isExpanded=e,this.render(),this.setupEventListeners()}destroy(){this.cleanupListeners(),this.container.innerHTML=""}}function bo(t){t.innerHTML=`
-    <h2>Sitemap Scan</h2>
-    <div id="view-toggle">
-      <button id="treeViewBtn" class="view-btn active">Tree View</button>
-      <button id="diagramViewBtn" class="view-btn">Diagram View</button>
-    </div>
-    <div id="filters-container">
-      <div id="filters"></div>
-    </div>
-    <div id="stats">
-      <p>Total URLs: <span id="totalUrls">0</span></p>
-      <p>Unique Domains: <span id="uniqueDomains">0</span></p>
-      <p>Matches: <span id="matchCount"></span></p>
-    </div>
-    <div id="view-container">
-      <div id="tree" class="tree"></div>
-      <div id="diagram" class="diagram view-hidden"></div>
-    </div>
-    <div id="status">
-      <div id="loading" class="hidden">
-        <progress id="loadProgress" max="100" value="0"></progress>
-        <span>Loading...</span>
-      </div>
-      <div id="feedback" class="hidden"></div>
-      <div id="error" class="hidden"></div>
-    </div>
-  `;const n=t.querySelector("#filters-container").querySelector("#filters"),i=t.querySelector("#tree"),r=t.querySelector("#diagram"),s=new yn(i),o=new vo(r);let a=[],u={},c="url",l="",f="tree";const d=new _o(n,{onApply:(b,C)=>{u=b,c=C,m(),v()},onReset:()=>{u={},c="url",l="",m(),v()},onSearch:b=>{l=b,m(),v()}});function m(){f==="tree"?(s.render(a,u,c,l),i.classList.remove("view-hidden"),r.classList.add("view-hidden")):(o.render(a,u,c,l),r.classList.remove("view-hidden"),i.classList.add("view-hidden"))}function v(){const C=(f==="tree"?i:r).querySelectorAll(f==="tree"?".url-leaf":".node text.url-label").length;t.querySelector("#matchCount").textContent=`(${C} of ${a.length} URLs)`}const _=t.querySelector("#treeViewBtn"),S=t.querySelector("#diagramViewBtn");return _.addEventListener("click",()=>{f!=="tree"&&(f="tree",_.classList.add("active"),S.classList.remove("active"),m(),v())}),S.addEventListener("click",()=>{f!=="diagram"&&(f="diagram",S.classList.add("active"),_.classList.remove("active"),m(),v())}),(async()=>{const C=new URLSearchParams(window.location.search).get("id");if(!C){et("No sitemap ID provided. Please load a sitemap first.");return}U(10);try{const $=await fetch(`/sitemap/results-data?id=${C}`);if(!$.ok){const N=await $.json();throw new Error(N.error||"Failed to load sitemap data")}const z=await $.json();a=z.urls,U(50),m(),U(100),Rt(`Loaded ${z.urlCount} URLs`),t.querySelector("#totalUrls").textContent=z.urlCount,t.querySelector("#uniqueDomains").textContent=new Set(a.map(N=>new URL(N.loc).hostname)).size,ot.updateUrls(a),v()}catch($){et($.message)}finally{K()}})(),{treeView:s,onToggleFilters:b=>{d.toggleVisibility(b)},urls:a}}let ot,B,tt=null;function Ot(t){if((!B||!document.getElementById("view"))&&(console.error("Main element not available or invalid for routing:",{mainElExists:!!B,viewInDOM:!!document.getElementById("view")}),B=document.getElementById("view"),!B))return;console.debug("Routing to:",t),tt&&(console.debug("Cleaning up previous page before routing"),tt(),tt=null),B.innerHTML="",console.debug("Cleared mainEl content");let e={};try{switch(t){case"/":console.debug("Setting up input page"),e=mn(B)||{},tt=null;break;case"/results":console.debug("Setting up results page"),e=bo(B)||{},tt=e.cleanup||null;break;default:B.innerHTML="<h2>404 - Page not found</h2>",tt=null}ot?(console.debug("Updating header with path:",t),ot.updateRoute(t),e.urls&&ot.updateUrls(e.urls),(e.treeView||e.onToggleFilters)&&ot.updateOptions({treeView:e.treeView,onToggleFilters:e.onToggleFilters})):console.warn("Header instance not initialized during route update"),console.debug("Route completed, mainEl content:",B.innerHTML.substring(0,100))}catch(n){console.error("Routing error:",n),B.innerHTML=`<h2>Error: ${n.message}</h2>`,tt=null}}document.addEventListener("DOMContentLoaded",()=>{B=document.getElementById("view");const t=document.getElementById("header");if(!B||!t){console.error("DOM initialization failed:",{mainFound:!!B,headerFound:!!t,documentBody:document.body.innerHTML.substring(0,200)}),document.body.innerHTML="<h2>Error: Required DOM elements not found</h2>";return}ot=new dn(t),Ot(window.location.pathname),window.addEventListener("popstate",e=>{console.debug("Popstate event triggered, routing to:",window.location.pathname),Ot(window.location.pathname)})});
