+++ public/index.html
@@ -5,8 +5,8 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <title>Sitemap Explorer</title>
   <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.1/css/all.min.css" integrity="sha512-DTOQO9RWCH3ppGqcWaEA1BIZOC6xxalwEsw9c2QQeAIftl+Vegovlnee1c9QX4TctnWMn13TZye+giMm8e2LwA==" crossorigin="anonymous" referrerpolicy="no-referrer" />
-  <script type="module" crossorigin src="/assets/index-Bh74FOVn.js"></script>
-  <link rel="stylesheet" crossorigin href="/assets/index-D0sDutv4.css">
+  <script type="module" crossorigin src="/assets/index-DfX2DsY0.js"></script>
+  <link rel="stylesheet" crossorigin href="/assets/index-DrZbJhIi.css">
 </head>
 <body>
   <div id="app">
