+++ src/client/components/TreeView/TreeView.js
@@ -1,5 +1,6 @@
 // src/client/components/TreeView.js
-import { showLoading, hideLoading } from '../utils/dom.js';
+import { showLoading, hideLoading } from '../../utils/dom.js';
+import './styles.css';
 
 function debounce(fn, delay) {
   let timeout;
