+++ src/client/components/TreeView/styles.css
@@ -1,0 +1,77 @@
+/* Tree View */
+.tree-controls {
+    margin-bottom: 10px;
+}
+
+.tree-list {
+    list-style: none;
+    padding-left: 0;
+}
+
+.tree-list ul {
+    padding-left: 20px;
+}
+
+.tree-list li {
+    margin: 6px 0;
+    display: inherit;
+}
+
+.toggle {
+    cursor: pointer;
+    user-select: none;
+    display: inline-flex;
+    align-items: center;
+    padding: 4px 8px;
+    background: #f1f1f1;
+    border-radius: 4px;
+    transition: background 0.2s ease;
+}
+
+.dark-mode .toggle {
+    background: #3a3a3a;
+}
+
+.toggle:hover {
+    background: #e0e0e0;
+}
+
+.dark-mode .toggle:hover {
+    background: #4a4a4a;
+}
+
+.toggle::before {
+    content: '▶';
+    margin-right: 6px;
+    color: #007bff;
+    font-size: 0.8em;
+    transition: transform 0.2s ease;
+}
+
+.dark-mode .toggle::before {
+    color: #1e90ff;
+}
+
+.toggle.open::before {
+    content: '▼';
+    transform: rotate(0deg);
+}
+
+.url-leaf {
+    padding: 4px 8px;
+    display: block;
+    color: #0066cc;
+}
+
+.dark-mode .url-leaf {
+    color: #1e90ff;
+}
+
+.highlight {
+    background: #fff3cd;
+}
+
+.dark-mode .highlight {
+    background: #ffd700;
+    color: #2d2d2d;
+}
